% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.files.to.emuDB.R
\name{add.files.to.bundles}
\alias{add.files.to.bundles}
\title{Add files to bundles}
\usage{
add.files.to.bundles(path2rootDir, path2sessionDir, fileExt,
  mediafileExtension = "wav")
}
\arguments{
\item{path2rootDir}{path to directory containing files that should be added}

\item{path2sessionDir}{path to session directory containing the destination bundles}

\item{fileExt}{file extention of the files that should be added to the bundles}

\item{mediafileExtension}{file extention of media files of emuDB (default='wav')}
}
\description{
Matches the basenames of all files found in path2rootDir
with the according mediafiles contained in the session
folder specified and copies the files into the according
_bndl folders.
}
\author{
Raphael Winkelmann
}

