% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_attrDefLabelGroup}
\alias{add_attrDefLabelGroup}
\title{Add labelGroup to attributeDefinition of emuDB}
\usage{
add_attrDefLabelGroup(dbName, levelName, attributeDefinitionName,
  labelGroupName, labelGroupValues, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{labelGroupName}{name of label group}

\item{labelGroupValues}{character vector of labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Adds labelGroup that can be used as a short hand  
to reference groups of labels specific
to an attribute definition (compared to global label groups that 
are added by \code{\link{add_labelGroup}}) in a 
\code{\link{query}}. A common example would be to
add a label group for something like the phonetic
category of nasals to be able reference them 
as "nasals" in a \code{\link{query}}. For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

sampaNasals = c("m", "F", "n", "J", "N")

# add these values to the default Phonetic attribute
# definition of the Phonetic level of the "ae" emuDB
add_attrDefLabelGroup(dbName = "ae",
                      levelName = "Phonetic",
                      attributeDefinitionName = "Phonetic",
                      labelGroupName = "sampaNasals",
                      labelGroupValues = sampaNasals)

# query the labelGroup
query("ae", "Phonetic=sampaNasals")

}
}
\seealso{
list_attrDefLabelGroups, add_labelGroup
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

