% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_attributeDefinition}
\alias{add_attributeDefinition}
\title{Add attribute definition to emuDB}
\usage{
add_attributeDefinition(dbName, levelName, name, type = "STRING",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{name}{name af new attributeDefinition}

\item{type}{type of new attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add additional attribute definition to an existing level definition
of a emuDB. Each level definition is required to have at least one 
default attribute definition that has the same name as the level definition
(automatically created by \code{\link{add_levelDefinition}}). Attribute
definitions can be viewed as definitions of
parallel labels for the annotational units (ITEMs) of the emuDB. For more 
information on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add additional attribute definition to the "Phonetic" level
# of the "ae" emuDB that will contain the UTF8 IPA
# symbols of the phonetic transrictions
add_attributeDefinition(dbName = "ae",
                        levelName = "Phonetic",
                        name = "IPA-UTF8")

}
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

