% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.files.R
\name{add_files}
\alias{add_files}
\title{Add files to emuDB}
\usage{
add_files(dbName, dir, fileExtension, targetSessionName = "0000",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{dir}{directory containing files to be added}

\item{fileExtension}{file extension of file to be added}

\item{targetSessionName}{name of sessions containing 
bundles that the files will be added to}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add files to bundles of specified session of emuDB.
The files that are found in dir that have the extension 
fileExtension will be copied into the the according bundle
folder that have the same basename as the according file.
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

