% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_labelGroup}
\alias{add_labelGroup}
\title{Add (global) labelGroup to emuDB}
\usage{
add_labelGroup(dbName, name, values, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of label group}

\item{values}{character vector of labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Adds labelGroup that can be used as a short hand  
to reference groups of labels that are globally defined
for the entire database (compared to attributeDefinition
specific label groups that 
are added by \code{\link{add_attrDefLabelGroup}}) in a 
\code{\link{query}}. A common example would be to
add a label group for something like the phonetic
category of nasals to be able to reference them 
as "nasals" in a \code{\link{query}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

sampaNasals = c("m", "F", "n", "J", "N")

# add these values to the to the "ae" emuDB
# as a globally available labelGroup
add_labelGroup(dbName = "ae",
               name = "sampaNasals",
               values = sampaNasals)

# query the labelGroup in the "Phonetic" level
query("ae", "Phonetic == sampaNasals")

# query the labelGroup in the "Phoneme" level
query("ae", "Phoneme == sampaNasals")

}

}
\author{
Raphael Winkelmann
}
\seealso{
list_labelGroups, add_attrDefLabelGroup
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

