% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_linkDefinition}
\alias{add_linkDefinition}
\title{Add linkDefinition to emuDB}
\usage{
add_linkDefinition(dbName, type, superlevelName, sublevelName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{type}{type of linkDefinition}

\item{superlevelName}{name of superlevel of linkDefinition}

\item{sublevelName}{name of sublevel of linkDefinition}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Add linkDefinition to emuDB
}
\author{
Raphael Winkelmann
}

