% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{add_perspective}
\alias{add_perspective}
\title{Add perspective to emuDB}
\usage{
add_perspective(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of perspective}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add EMUwebAppConfig$perspective to emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

