% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_ssffTrackDefinitions}
\alias{add_ssffTrackDefinitions}
\title{Add ssffTrackDefinitions to emuDB}
\usage{
add_ssffTrackDefinitions(dbName = NULL, name = NULL, columnName = NULL,
  fileExtension = NULL, onTheFlyFunctionName = NULL,
  onTheFlyParams = NULL, onTheFlyOptLogFilePath = NULL, dbUUID = NULL,
  showProgress = TRUE, interactive = TRUE)
}
\arguments{
\item{dbName}{name of emuDB}

\item{name}{name of ssffTrackDefinitions}

\item{columnName}{columnName of}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Add ssffTrackDefinitions to emuDB
}

