% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{AddListRemoveLabelGroup}
\alias{AddListRemoveLabelGroup}
\alias{add_labelGroup}
\alias{list_labelGroups}
\alias{remove_labelGroup}
\title{Add / List / Remove global labelGroup to / of / from emuDB}
\usage{
add_labelGroup(dbName, name, values, dbUUID = NULL)

list_labelGroups(dbName, dbUUID = NULL)

remove_labelGroup(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of label group}

\item{values}{character vector of labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add / List / Remove label group that can be used as a short hand  
to reference groups of labels that are globally defined
for the entire database (compared to attribute definition
specific label groups that 
are added by \code{\link{add_attrDefLabelGroup}}) in a 
\code{\link{query}}. A common example would be to
add a label group for something like the phonetic
category of nasals to be able to reference them 
as "nasals" in a \code{\link{query}}. For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

sampaNasals = c("m", "F", "n", "J", "N")

# add these values to the "ae" emuDB
# as a globally available labelGroup
add_labelGroup(dbName = "ae",
               name = "sampaNasals",
               values = sampaNasals)

# query the labelGroup in the "Phonetic" level
query(dbName = "ae", 
      query = "Phonetic == sampaNasals")

# query the labelGroup in the "Phoneme" level
query(dbName = "ae", 
      query = "Phoneme == sampaNasals")

# list global label groups of "ae" emuDB
list_labelGroups(dbName = "ae")

# remove the newly added labelGroup
remove_labelGroup(dbName = "ae",
                  name = "sampaNasals")
}

}
\seealso{
add_attrDefLabelGroup
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

