% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{AddListRemoveLevelDefinitions}
\alias{AddListRemoveLevelDefinitions}
\alias{add_levelDefinition}
\alias{list_levelDefinitions}
\alias{remove_levelDefinition}
\title{Add / List / Remove level definition to / of / from emuDB}
\usage{
add_levelDefinition(dbName, name, type, store = TRUE, dbUUID = NULL)

list_levelDefinitions(dbName, dbUUID = NULL)

remove_levelDefinition(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of level definition}

\item{type}{type of level definition ("SEGMENT","EVENT","ITEM")}

\item{store}{changes to file system (TRUE), changes only in memory (FALSE)}

\item{dbUUID}{optional UUID of loaded emuDB, in case dbName is ambique}
}
\description{
Add / List / Remove database operation functions for level definitions. 
A level is a more general term for what is often referred to as a "tier". 
It is more general in the sense that people usually 
expect tiers to contain time information. Levels 
can either contain time information if they are of the 
type "EVENT" or of the type "SEGMENT" but are timeless 
if they are of the type "ITEM". For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
Note that a level cannot be removed, if it contains instances of annotation items
or if it is linked to another level.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add level called "Phonetic2" to the "ae" emuDB
# that could for example contain the transcriptions of a second annotator
add_levelDefinition(dbName = "ae",
                    name = "Phonetic2",
                    type = "SEGMENT")
                    
# list level definition of "ae" emuDB
list_levelDefinitions(dbName = "ae")

# remove newly added level definition
remove_levelDefinitions(dbName = "ae",
                        name = "Phonetic2")
}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

