% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autobuild.R
\name{autobuild.linkFromTimes}
\alias{autobuild.linkFromTimes}
\title{Autobuild links between two levels using their time information}
\usage{
autobuild.linkFromTimes(db, superlevelName, sublevelName, writeToDisc = TRUE)
}
\arguments{
\item{db}{emuDB object to perform autobuild on}

\item{superlevelName}{name of level to link from (link definition required in db)}

\item{sublevelName}{name of level to link to (link definition required in db)}

\item{writeToDisc}{should changes be written to disk completing after autobuild}
}
\value{
an emuDB object
}
\description{
Autobuild links between two time levels. The super-level has to be of the
type SEGMENT and the sub-level either of type EVENT or of type SEGMENT. If
this is the case and a according link definition is present in db$DBconfig$linkDefintions,
this function automatically links the events or segments of the sub-level which occur
within (startSample to (startSample + sampleDur)) the segments of the super-level to those segments.
The linkDefinition$type (ONE_TO_MANY, MANY_TO_MANY, ONE_TO_ONE) is relavant whether a link
is genarated or not. See the emuR_autobuild vignette for more details.
}
\author{
Raphael Winkelmann
}
\keyword{autobuild}
\keyword{emuR}

