% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobuild.R
\name{autobuild_linkFromTimes}
\alias{autobuild_linkFromTimes}
\title{Autobuild links between two levels using their time information}
\usage{
autobuild_linkFromTimes(dbName, superlevelName, sublevelName,
  writeToFS = TRUE, convertSuperlevel = FALSE,
  backupLevelAppendStr = "-autobuildBackup", dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{superlevelName}{name of level to link from (link definition required in emuDB)}

\item{sublevelName}{name of level to link to (link definition required in emuDB)}

\item{writeToFS}{should changes be written to file system (annotJSON files) after completing autobuild process.}

\item{convertSuperlevel}{if set to TRUE a backup of the superlevel will be created and the actural
superlevel will be converted to a level of type ITEM}

\item{backupLevelAppendStr}{string appended to level name for backup level}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Autobuild links between two time levels. The super-level has to be of the 
type SEGMENT and the sub-level either of type EVENT or of type SEGMENT. If
this is the case and a according link definition is defined for the emuDB,
this function automatically links the events or segments of the sub-level which occur
within (startSample to (startSample + sampleDur)) the segments of the super-level to those segments.
The type of linkDefinition (ONE_TO_MANY, MANY_TO_MANY, ONE_TO_ONE) is relevant whether a link
is genarated or not (e.g. overlapping segments are linked in a MANY_TO_MANY relationship 
but not in a ONE_TO_MANY relationship). For more information on the structural 
elements of an emuDB see \code{vignette(emuDB)}.
}
\keyword{"ONE_TO_MANY",}
\keyword{"Phoneme")}
\keyword{"Phoneme",}
\keyword{"Syllable",}
\keyword{"myTGcolDB"}
\keyword{"myTGcolDB",}
\keyword{#}
\keyword{##################################}
\keyword{##'}
\keyword{(see}
\keyword{=}
\keyword{?load_emuDB}
\keyword{@examples}
\keyword{TRUE)}
\keyword{TextGridCollection}
\keyword{\code{vignette(emuR_intro)}}
\keyword{\dontrun{}
\keyword{add}
\keyword{add_linkDefinition(dbName}
\keyword{and}
\keyword{as}
\keyword{autobuild}
\keyword{autobuild_linkFromTimes(dbName}
\keyword{be}
\keyword{build}
\keyword{convertSuperlevel}
\keyword{converted}
\keyword{emuDB}
\keyword{emuR}
\keyword{envoke}
\keyword{for}
\keyword{function}
\keyword{has}
\keyword{hierarchy}
\keyword{information)}
\keyword{linkDefinition}
\keyword{loaded}
\keyword{more}
\keyword{one}
\keyword{prerequisite:}
\keyword{present}
\keyword{sublevelName}
\keyword{superlevelName}
\keyword{the}
\keyword{to}
\keyword{type}
\keyword{work}
\keyword{}}

