% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayesdist.R
\name{bayesplot}
\alias{bayesplot}
\title{bayesplot}
\usage{
bayesplot(data, train, N = 10, ellipse = FALSE, labs = NULL, xlab = "",
  ylab = "", colour = TRUE, ...)
}
\description{
bayesplot
}
\keyword{internal}

