% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.trackdata}
\alias{bind.trackdata}
\title{bind trackdata}
\usage{
\method{bind}{trackdata}(...)
}
\arguments{
\item{\dots}{trackdata objects}
}
\description{
binds diffrent trackdata objects together
}
\keyword{methods}

