\name{bridge}
\alias{bridge}
\title{Emu segment list}
\usage{data(bridge)}
\description{
   A matrix of the (hypothetical) count per hour of cars crossing a bridge on three separate days.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}