% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.database.R
\name{contextQuery}
\alias{contextQuery}
\title{Query emuR query result context}
\usage{
contextQuery(seglist, offset = 0, offsetRef = "START", seqLength = 1,
  resultType = NULL, dbUUID = NULL)
}
\arguments{
\item{resultType}{type (class name) of result}

\item{dbObj}{object of class emuDB}

\item{segs}{segment list}

\item{level}{level of input segment list}

\item{targetlevel}{target level}

\item{justlabels}{labels only}

\item{sequence}{integer which determins sibling position relative to segment in segs}

\item{longerok}{TODO ??}
}
\value{
result set object of class resultType (e.g. EMU seglist 'emusegs')
}
\description{
Requery an EMU database
}
\examples{
\dontrun{
## Requery phonetic context of syllable

requery(ae,sl1,level='Syllable',targetlevel='Phonetic')

## Requery previous element on phonetic level

requery(ae,sl2,level='Phonetic',sequence=-1)

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

