% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.database.R
\name{contextRequery}
\alias{contextRequery}
\title{Query emuR result context}
\usage{
contextRequery(seglist, offset = 0, offsetRef = "START", seqLength = 1,
  resultType = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list}

\item{offset}{offset in sequence}

\item{offsetRef}{referenec of offset: 'START' start of segments, 'END': end of segments}

\item{seqLength}{item length of new segment list}

\item{resultType}{type (class name) of result}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class resultType (default: 'emuRsegs')
}
\description{
Requery an EMU database
}
\examples{
\dontrun{

## Requery previous element of 'p' on phonetic level
sl1=query('ae','Phonetic=p')
contextRequery(sl1,offset=-1)

## Requery context (previuos and following) of 'p' on phonetic level

contextRequery(sl1,offset=-1,seqLength=3)



}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{EQL2}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

