% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.database.from.legacy.emu.by.name}
\alias{convert.database.from.legacy.emu.by.name}
\title{Convert legacy EMU database and store in new format to directory
Loading by name only works if databse was used with legacy EMU.}
\usage{
convert.database.from.legacy.emu.by.name(dbName, targetDir,
  showProgress = TRUE)
}
\arguments{
\item{dbName}{legacy EMU database name}

\item{targetDir}{target directory}

\item{showProgress}{show progress bar}
}
\value{
object of class emuSX.database
}
\description{
Convert legacy EMU database and store in new format to directory
Loading by name only works if databse was used with legacy EMU.
}
\examples{
\dontrun{
## Load database "ae", convert and save in new format to directory /homes/mylogin/EMUnew/

ae=convert.database.from.legacy.emu.by.name("ae","/homes/mylogin/EMUnew/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{convert.database.from.legacy.emu}} \code{\link{load.database}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuSX}
\keyword{schema}

