% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.database.from.legacy.emu}
\alias{convert.database.from.legacy.emu}
\title{Convert legacy EMU database and store in new format to directory}
\usage{
convert.database.from.legacy.emu(emuTplPath, targetDir, showProgress = TRUE)
}
\arguments{
\item{emuTplPath}{EMU template file path}

\item{targetDir}{target directory}

\item{showProgress}{show progress bar}
}
\value{
object of class emuR.database
}
\description{
Convert legacy EMU database and store in new format to directory
}
\examples{
\dontrun{
## Convert legacy EMU database specified by EMU template file /homes/mylogin/ae/ae.tpl to directory /homes/mylogin/EMUnew/ae

convert.database.from.legacy.emu("/homes/mylogin/ae/ae.tpl","/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.database}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuR}
\keyword{schema}

