% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legacy.R
\name{convert.legacyEmuDB.by.name.to.emuDB}
\alias{convert.legacyEmuDB.by.name.to.emuDB}
\title{Convert legacy EMU database and store it in new format
If the legacy database could be found it is loaded. If load is successfull a new directory with the name of the database is created in the \code{targetDirectory}
Loading by name only works if database was used with legacy EMU. Use the function \code{\link{convert.legacyEmuDB.to.emuDB}} otherwise.}
\usage{
convert.legacyEmuDB.by.name.to.emuDB(dbName, targetDir, options = NULL,
  verbose = TRUE)
}
\arguments{
\item{dbName}{legacy EMU database name}

\item{targetDir}{target directory}

\item{options}{list of options}

\item{verbose}{be verbose}
}
\description{
options is a list of key value pairs:
rewriteSSFFTracks if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: TRUE
ignoreMissingSSFFTrackFiles if TRUE missing SSFF track files are ignored, default: TRUE
##' sourceFileTextEncoding encoding of legacy database text files (template, label and hlb files) :default NULL (usess encoding of operating system platform)
}
\examples{
\dontrun{
## Load database "ae", convert and save in new format to directory /homes/mylogin/EMUnew/

convert.legacyEmuDB.by.name.to.emuDB("ae","/homes/mylogin/EMUnew/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{convert.legacyEmuDB.to.emuDB}} \code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

