% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legacy.R
\name{convert.legacyEmuDB.to.emuDB}
\alias{convert.legacyEmuDB.to.emuDB}
\title{Convert legacy EMU database and store it in new format}
\usage{
convert.legacyEmuDB.to.emuDB(emuTplPath, targetDir, options = NULL,
  verbose = TRUE)
}
\arguments{
\item{emuTplPath}{EMU template file path}

\item{targetDir}{target directory}

\item{options}{list of options}

\item{verbose}{be verbose}
}
\description{
If the legacy database template file could be found the database metadata and annoations are loaded. If load is successfull a new directory with the name of the database is created in the \code{targetDirectory}.
}
\details{
Information of the legacy Emu template file is transferred to [dbname]_DBconfig.json file. Legacy Emu utterances are reorganized in sessions and bundles.
Media files (e.g. wav files) are copied, SSFF track files are rewritten. Annotations in Emu hierarchy (.hlb) files and ESPS label files are converted to a [bundleName]_annot.json file per bundle (utterance).
Please note that only those files get copied, which are referenced by the template file. Additional files in the legacy database directories are ignored. The legacy Emu database is not modified.

options is a list of key value pairs:
rewriteSSFFTracks if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: TRUE
ignoreMissingSSFFTrackFiles if TRUE missing SSFF track files are ignored, default: TRUE
}
\examples{
\dontrun{
## Convert legacy EMU database specified by EMU
## template file /homes/mylogin/ae/ae.tpl to directory /homes/mylogin/EMUnew/ae

convert.legacyEmuDB.to.emuDB("/homes/mylogin/ae/ae.tpl","/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

