% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.TextGridCollection.to.emuDB}
\alias{convert.TextGridCollection.to.emuDB}
\title{Convert a TextGridCollection (.wav & .TextGrid files) to a emuDB}
\usage{
convert.TextGridCollection.to.emuDB(path2rootDir, dbName, targetDir,
  tgExt = "TextGrid", audioExt = "wav", showProgress = TRUE)
}
\arguments{
\item{path2rootDir}{path to root directory (CAUTION: think of DB size and search space!)}

\item{dbName}{name given to newly generated emuDB}

\item{targetDir}{directory where to save newly generated emuDB}

\item{tgExt}{extention of TextGrid files (default=TextGrid meaning filesnames of the form baseName.TextGrid)}

\item{audioExt}{extention of audio files (default=wav meaning filesnames of the form baseName.wav).}

\item{showProgress}{show progress bar flag}
}
\description{
Converts a TextGridCollection to a emuDB by first generating a file pair list
containing the paths to the .wav & .TextGrid (default extentions) files with the same base
name. It then generates a emuDB DBconfig from the first TextGrid in this file pair list.
After this it converts all file pairs to the new format.
}
\author{
Raphael Winkelmann
}
\seealso{
create.filePairList
}

