% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{convert_legacyEmuDB_by_name_to_emuDB}
\alias{convert_legacyEmuDB_by_name_to_emuDB}
\title{Convert a known legacy EMU database to emuDB format}
\usage{
convert_legacyEmuDB_by_name_to_emuDB(dbName, targetDir, options = NULL,
  verbose = TRUE)
}
\arguments{
\item{dbName}{legacy EMU database name}

\item{targetDir}{target directory}

\item{options}{list of options (see details of \code{\link{convert_legacyEmuDB_to_emuDB}})}

\item{verbose}{be verbose}
}
\description{
Converts an existing known legacy EMU database to emuDB format. The legacy database is addressed by its name. Creates the target directory for the new structure, converts the database configuration, copies/rewrites signal files and converts annotations.
In the second step the target directory for the new database structure is created, the configuration and annoation files are stored in the new format and the signal files are copied.
}
\details{
This function first tries to get the path of the legacy template file from the given database name using \code{\link{list_legacyEmuDBs}}.
If the database could be found the function \code{\link{convert_legacyEmuDB_to_emuDB}} is called.
}
\examples{
\dontrun{
## Load database "ae", convert and save in new format to directory /homes/mylogin/EMUnew/

convert_legacyEmuDB_by_name_to_emuDB("ae","/homes/mylogin/EMUnew/")

}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{convert_legacyEmuDB_to_emuDB}} \code{\link{load_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

