% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{convert_legacyEmuDB_to_emuDB}
\alias{convert_legacyEmuDB_to_emuDB}
\title{Convert legacy EMU database to emuDB format}
\usage{
convert_legacyEmuDB_to_emuDB(emuTplPath, targetDir, dbUUID = NULL,
  options = NULL, verbose = TRUE)
}
\arguments{
\item{emuTplPath}{EMU template file path}

\item{targetDir}{target directory}

\item{dbUUID}{optional UUID of emuDB}

\item{options}{list of options (see details)}

\item{verbose}{be verbose}
}
\description{
Converts an existing legacy EMU database to emuDB format. Creates the target directory for the new structure, converts the database configuration, copies/rewrites signal files and converts annotations.
}
\details{
The database will be converted if the legacy database template file could be found and succesfully be loaded and parsed. A new directory with the name of the database will be created in the \code{targetDirectory}. \code{targetDirectory} and its parent directories will be created if the do not exist. If the new database directory already exists the function stops with an error.

Information of the legacy Emu template file is transferred to [dbname]_DBconfig.json file. Some of the flags of the legacy EMU template files are ignored (lines of syntax: "set [flagName] [flagValue]", known ignored flag names are: 'LabelTracks','SpectrogramWhiteLevel','HierarchyViewLevels','SignalViewLevels'). 
Legacy Emu utterances are reorganized to sessions and bundles. Session structure depends on wilcard path pattern of primary track. At least one default session with name '0000' will be created.
Media files (e.g. wav files) are copied, SSFF track files are rewritten (read/write with ASSP (wrassp) library) by default. Annotations in Emu hierarchy (.hlb) files and ESPS label files are converted to a [bundleName]_annot.json file per bundle (utterance).
Please note that only those files get copied, which are referenced by the template file. Additional files in the legacy database directories are ignored. The legacy Emu database will not be modified.

options is a list of key value pairs:

\code{rewriteSSFFTracks} if \code{TRUE} rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: \code{TRUE}

\code{ignoreMissingSSFFTrackFiles} if \code{TRUE} missing SSFF track files are ignored, default: \code{TRUE}

\code{sourceFileTextEncoding} encoding of legacy database text files (template, label and hlb files), possible values: NULL, "latin1", "UTF-8" "bytes" or "unknown" :default \code{NULL} (uses encoding of operating system platform)

\code{symbolicLinkSignalFiles} if \code{TRUE} signal files are symbolic linked instead of copied. Implies: \code{rewriteSSFFTracks=FALSE}, Default: \code{FALSE}
}
\examples{
\dontrun{
## Convert legacy EMU database specified by EMU 
## template file /homes/mylogin/ae/ae.tpl to directory /homes/mylogin/EMUnew/
## and load it afterwards

convert_legacyEmuDB_to_emuDB("/homes/mylogin/ae/ae.tpl","/homes/mylogin/EMUnew/")
ae=load_emuDB("/homes/mylogin/EMUnew/ae")

## Convert database "ae" and do not rewrite SSFF tracks 

convert_legacyEmuDB_by_name_to_emuDB(
"ae",
"/homes/mylogin/EMUnew/",
options=list(rewriteSSFFTracks=FALSE)
)

}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

