\name{coutts.epg}
\alias{coutts.epg}
\title{EPG-compressed trackdata of the segment list coutts}
\usage{data(coutts.epg)}
\description{
   EPG-compressed trackdata of the segment list coutts(frame rate 5 ms).
} 
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}