% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{coutts}
\alias{coutts}
\title{Segment list of words, read speech, female speaker of Australian English
from database epgcoutts}
\format{segmentlist}
\description{
An EMU dataset
}
\keyword{datasets}

