% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{coutts.sam}
\alias{coutts.sam}
\title{Trackdata of acoustic waveforms from the segment list coutts}
\format{trackdata object}
\description{
An EMU dataset
}
\keyword{datasets}

