\name{coutts2.epg}
\alias{coutts2.epg}
\title{EPG-compressed trackdata of the segment list coutts2}
\usage{data(coutts2.epg)}
\description{
   EPG-compressed trackdata of the segment list coutts2 (frame rate 5 ms).
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}