\name{coutts2.l}
\alias{coutts2.l}
\title{Vector of word label of the segment list coutts2}
\usage{data(coutts2.l)}
\description{
   Vector of word label of the segment list coutts2.
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}