\name{coutts2.sam}
\alias{coutts2.sam}
\title{Trackdata of acoustic waveforms of segment list coutts2}
\usage{data(coutts2.sam)}
\description{
   Trackdata of acoustic waveforms of coutts2 
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}