% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.database.schema.from.TextGrid}
\alias{create.database.schema.from.TextGrid}
\title{Create emuDB database schema object from a TextGrid file}
\usage{
create.database.schema.from.TextGrid(tgPath, dbName)
}
\arguments{
\item{tgPath}{path to TextGrid file}

\item{dbName}{name of the database}
}
\value{
object of class emuDB.schema.db
}
\description{
Create emuDB database schema object from a TextGrid file
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{TextGrid}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

