% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.DBconfig.from.TextGrid.R
\name{create.DBconfig.from.TextGrid}
\alias{create.DBconfig.from.TextGrid}
\title{Create emuDB DBconfig object from a TextGrid file}
\usage{
create.DBconfig.from.TextGrid(tgPath, dbName, tierNames = NULL)
}
\arguments{
\item{tgPath}{path to TextGrid file}

\item{dbName}{name of the database}

\item{tierNames}{character vector containing names of tiers to extract and convert. If NULL (the default) all
tiers are converted.}
}
\value{
object of class emuDB.schema.db
}
\description{
Create emuDB DBconfig object from a TextGrid file
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{TextGrid}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

