% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_demoData.R
\name{create_demoData}
\alias{create_demoData}
\title{Create demo data for emuR package}
\usage{
create_demoData(dir = tempdir())
}
\description{
Will create a folder in the folder specified
by the dir argument called emuR_demoData.
This folder contains the folders:
\itemize{
\item{ae: }{Containing an emuDB that adheres to the new formant specification (as expected by the \code{load_emuDB()} function)}
\item{TextGrid_collection: }{Containing a TextGrid collection (as expected from the \code{convert_TextGridCollection_to_emuDB()} function)}
}
}

