% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{create_emuDB}
\alias{create_emuDB}
\title{Create empty emuDB}
\usage{
create_emuDB(name, targetDir, mediaFileExtension = "wav")
}
\arguments{
\item{name}{of new emuDB}

\item{targetDir}{target directory in which to store the emuDB}

\item{mediaFileExtension}{defines mediaFileExtention (NOTE: currently only
'wav' (the default) is supported by all components of EMU)}
}
\description{
creates an empty emuDB in the target directory specified
}
\author{
Klaus Jaensch
}

