% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.filePairList.R
\name{create_filePairList}
\alias{create_filePairList}
\title{Create a file-pair-list}
\usage{
create_filePairList(ext1Path2rootDir, ext2Path2rootDir, ext1, ext2)
}
\arguments{
\item{ext1Path2rootDir}{path to root directory of first file extention}

\item{ext2Path2rootDir}{path to root directory of second file extention (CAUTION: think of DB size and search space!)}

\item{ext1}{first extention to look for. This extention is considered the primary extention.
This means that this extentions genarates the basename list that the second extentions list is matched against.}

\item{ext2}{second extention to pair base names of first extention with}
}
\description{
Recursivly searches through a root directory and matches the
basenames of files that have the extentions provided.
}

