% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-annotations_crud.R
\name{create_links}
\alias{create_links}
\title{create links between items}
\usage{
create_links(emuDBhandle, links, rewriteAllAnnots = TRUE,
  verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{links}{data.frame containing linking information. The required columns
are: 
\itemize{
\item \code{session}: 
\item \code{bundle}
\item \code{from_id}
\item \code{to_id}
}}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json
files) (intended for expert use only)}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
create links between items
}
