% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuRtrackdata.R
\name{cut.emuRtrackdata}
\alias{cut.emuRtrackdata}
\title{Function to extract an emuRtrackdata at a single time point of to create another EMU-trackdata object between two times}
\usage{
\method{cut}{emuRtrackdata}(emuRtrackdata, leftTime, rightTime, single = TRUE,
  average = TRUE, prop = FALSE)
}
\arguments{
\item{emuRtrackdata}{emuRtrackdata object}

\item{leftTime}{bla}

\item{rightTime}{bli}

\item{single}{blup}

\item{average}{blap}

\item{prop}{woooord}
}
\description{
this function is intended to mimic some of the dcut behaviour
}

