\name{dbinfo}
\alias{dbinfo}  
\title{
Available EMU Speech databases
}
\description{
dbinfo lists all available databases (templates) in the linked Emu System
}
\usage{
dbinfo()
}
\value{
Returns the name of the template files without extension.
}
\author{
Tina John
}

\seealso{
\code{\link{trackinfo}}
}
\examples{
\dontrun{dbinfo()}
}
\keyword{utilities}
