% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlDBactions.R
\name{delete_unwanted_levels_from_database_tables}
\alias{delete_unwanted_levels_from_database_tables}
\title{Delete unwanted levels from database tables}
\usage{
delete_unwanted_levels_from_database_tables(conn, itemsTableName,
  labelsTableName, linksTableName, namesOfLevelsToKeep)
}
\arguments{
\item{conn}{sql DB connection}

\item{itemsTableName}{SQL table name of items table}

\item{linksTableName}{SQL table name of links table}

\item{namesOfLevelsToKeep}{character vector containing
the level names that should remain in the tables}

\item{labelTableName}{SQL table name of label table}
}
\description{
Deletes all item, label and link rows that do not
belong to the levels specified in namesOfLevelsToKeep
}
\author{
Raphael Winkelmann
}

