% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{demo.all.fm}
\alias{demo.all.fm}
\title{Formant track data for segment list demo.vowels}
\format{index: a two columned matrix with the range of the \$data rows that
belong to the segment ftime: a two columned matrix with the times marks of
the segment data: a three columned matrix with the formant values of the
first three formants for each segment}
\description{
A track list of the demo database that is part of the Emu system.  It is
the result of get fm data for the segment list demo.vowels (see
data(demo.vowels)).
}
\details{
A track list is created via emu.track() or via get data within the EMU
Query Tool.
}
\seealso{
\code{\link{demo.all.rms}} \code{\link{segmentlist}}
\code{\link{trackdata}}
}
\keyword{datasets}

