% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{demo.vowels}
\alias{demo.vowels}
\title{Emu segment List}
\format{First Column labels Second start time of the segment Third end time
of the segment Fourth utterance name of the utterance the segment was found}
\description{
Segment list of the demo database that is part of the Emu system.  It is
the result of a database query, that searched all vowel segments at level
Phonetic.
}
\details{
A segment list is created via \code{\link{query}}.
}
\seealso{
\code{\link{demo.all}} \code{\link{segmentlist}}
}
\keyword{datasets}

