% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dimnames.trackdata}
\alias{dimnames.trackdata}
\title{Dimnames of trackdata object}
\usage{
\method{dimnames}{trackdata}(x)
}
\arguments{
\item{x}{trackdata object}
}
\description{
returns dimension names of trackdata objects
}
\keyword{methods}

