\name{dip.l}
\alias{dip.l}
\title{Vector of phoneme labels from segment list dip}
\usage{data(dip)}
\description{
  For dip see data(dip)
}
%\format{
%  NA
%
%}
%
\details{ 
   dip  = emu.query("kielread06", "*", "[Kanonic=aI | aU | OY]")
   dip.l = label(dip)
}

\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}


\keyword{datasets}
\keyword{emu}