\name{dip}
\alias{dip}
\title{Emu segment list}
\usage{data(dip)}
\description{
   Segment list of the kielread06 database that is part of the Emu system.
   It is the result of a database query, that searched all three German diphthongs at level Kanonic.
}
%\format{
%  NA
%
%}
%
\details{ 
   dip  = emu.query("kielread06", "*", "[Kanonic=aI | aU | OY]")
}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}