\name{dorfric.dft}
\alias{dorfric.dft}
\title{Emu trackdata object}
\usage{data(dorfric.dft)}
\description{Spectral trackdata object, 256 point DFT, 16 kHz 
			sampling frequency.
   
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}