\name{dorfric.l}
\alias{dorfric.l}
\title{Vector of phoneme labels from segment list dorfric}
\usage{data(dorfric.l)}
\description{
   A vector of their labels: "C" or "x"
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}