\name{dorfric.lv}
\alias{dorfric.lv}
\title{Vector of preceding phoneme labels from segment list dorfric}}
\usage{data(dorfric.lv)}
\description{
   A vector of the preceding labels of the labels: "C" or "x"
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}