\name{dorfric.vl}
\alias{dorfric.vl}
\title{A vector of the preceding vowel labels in the dorfric dataset}
\usage{dorfric.vl}
\description{A vector of labels preceding the labels: "C" or "x"
   
}
%\format{
%  NA
%
%}
%


\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

%\keyword{emu}
\keyword{datasets}