\name{dorfric.w}
\alias{dorfric.w}
\title{A vector of the word labels for segment list dorfric}
\usage{data(dorfric.w)}
\description{
   A vector of the word labels for the segment list dorfric
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}