\name{dorsal.bound}
\alias{dorsal.bound}
\title{Vector of the acoustic boundary times between V and C in dorsal}
\usage{data(dorsal.bound)}
\description{
   Vector of the acoustic boundary times between V and C in dorsal.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}