\name{dorsal.clab}
\alias{dorsal.clab}
\title{A vector of consonant labels for C (dorsal)}
\usage{data(dorsal.clab)}
\description{
   A vector of consonant labels for C (dorsal).
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}