\name{dorsal.epg}
\alias{dorsal.epg}
\title{EPG-compressed trackdata of the segment list dorsal}
\usage{data(dorsal.epg)}
\description{
   EPG-compressed trackdata of the segment list dorsal.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}