\name{dorsal.sam}
\alias{dorsal.sam}
\title{Trackdata of acoustic waveform for the segment list dorsal}
\usage{data(dorsal.sam)}
\description{
   Trackdata of acoustic waveform for the segment list dorsal.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}