% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dplot.norm}
\alias{dplot.norm}
\title{dplot.norm}
\usage{
dplot.norm(x, labs = NULL, average = FALSE, xlim = NULL, ylim = NULL,
  lty = FALSE, type = "l", colour = TRUE, lwd = NULL, pch = NULL,
  legend = "topright", n = 20)
}
\description{
see function
}
\keyword{internal}

