% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dplot.time}
\alias{dplot.time}
\title{dplot.time}
\usage{
dplot.time(x, labs = NULL, offset = 0, prop = TRUE, average = FALSE,
  xlim = NULL, ylim = NULL, lty = FALSE, colour = TRUE, lwd = NULL,
  pch = NULL, legend = "topright", type = "l", ...)
}
\description{
see function
}
\keyword{internal}

