\name{dur.trackdata}
\alias{dur.trackdata}
\title{ Duration of trackdata elements}
\description{
Duration of segments is calculated for each element in the trackdata object
}
\usage{
\method{dur}{trackdata}(x)
}
\arguments{
  \item{x}{ a trackdata object }
}
\value{
a vector of durations
}
\author{Jonathan Harrington}
\keyword{ internal }
