% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{e.dft}
\alias{e.dft}
\title{Spectral vector of a single E vowel produced by a male speaker of Standard
North German.}
\format{spectral vector}
\description{
An EMU dataset
}
\keyword{datasets}

