\name{e.dft}
\alias{e.dft}
\title{Emu trackdata object}
\usage{data(e.dft)}
\description{
   a spectral vector of a German vowel [E] obtained with a DFT of 512 points at a sampling frequency of 16000 Hz.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}