\name{emu.defaultpaths}
\Rdversion{1.1}
\alias{emu.defaultpaths}
\title{
Default paths of the emu and tcl/tk libraries
}
\description{
 Returns the paths that are suggested for the system for the emu and tcl/tk libraries.
}
\usage{
emu.defaultpaths()
}

\author{
Tina John
}

\examples{
emu.defaultpaths()
}
\keyword{IO}
