\name{emu.variables}
\alias{emu.version}
\alias{emu.date}
\alias{emu.year}
\docType{data}
\title{ Emu variables}
\description{
  \describe{
    \item{\code{emu.version}}{Current version of the emu R package}
    \item{\code{emu.date}}{Date of package creation}
    \item{\code{emu.year}}{Year of package creation, copyright.}
  }
}
\usage{
emu.version
emu.date
emu.year
}
\keyword{datasets}
