\name{emuinfo}
\alias{emuinfo}  
\title{
Version number of the EMU tcl-library in tcl path
}
\description{
emuinfo returns the version number of the EMU tcl-library in tcl path
}
\usage{
emurinfo()
}
\value{
Returns the version number of the EMU tcl-library in tcl path.
}
\author{
Tina John
}

\seealso{
\code{\link{emurinfo}}
}
\examples{
\dontrun{emuinfo()}
}
\keyword{utilities}
