\name{emurinfo}
\alias{emurinfo}  
\title{
Version number of the EMU/R tcl-library in tcl path
}
\description{
emurinfo returns the version number of the EMU/R tcl-library in tcl path
}
\usage{
emurinfo()
}
\value{
Returns the nversion number of the EMU/R tcl-library in tcl path.
}
\author{
Tina John
}

\seealso{
\code{\link{emuinfo}}
}
\examples{
\dontrun{emurinfo()}
}
\keyword{utilities}
