\name{engassim.epg}
\alias{engassim.epg}
\title{EPG-compressed trackdata of the segment list engassim}
\usage{data(engassim.epg)}
\description{
   EPG-compressed trackdata of the segment list engassim.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}