% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{engassim.l}
\alias{engassim.l}
\title{Vector of phonetic labels from the segment list engassim: nK = nk,ng , sK =
sk,sg}
\format{vector of phonetic labels}
\description{
An EMU dataset
}
\keyword{datasets}

