\name{engassim}
\alias{engassim}
\title{Emu segment list}
\usage{data(engassim)}
\description{
   Segment list of sequences of /nk, ng, sk, sg/, one female speaker of Australian English.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}