% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRsegs.R
\name{export_seglistToTxtCollection}
\alias{export_seglistToTxtCollection}
\title{Exports a segment list to txt collection}
\usage{
export_seglistToTxtCollection(emuDBhandle, seglist, targetDir)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{seglist}{\code{tibble}, \code{\link{emuRsegs}} or 
\code{\link{emusegs}} object obtained by \code{\link{query}}ing a loaded emuDB}

\item{targetDir}{target directory to store}
}
\description{
Extract the media file (usually .wav file) snippets that correspond to 
the segments of a segment list (see result of a \code{\link{query}}) and 
save them to separate files and write the corresponding labels into a .txt file. Further,
the segmentlist is also stored to the target directory (as a .csv file).
}
