% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-export_TextGridCollection.R
\name{export_TextGridCollection}
\alias{export_TextGridCollection}
\title{Export to TextGridCollection}
\usage{
export_TextGridCollection(emuDBhandle, targetDir, sessionPattern = ".*",
  bundlePattern = ".*", attributeDefinitionNames = NULL)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{targetDir}{directory where the TextGrid collection should be saved}

\item{sessionPattern}{A regular expression pattern matching session names to be exported from the database}

\item{bundlePattern}{A regular expression pattern matching bundle names to be exported from the database}

\item{attributeDefinitionNames}{list of names of attributeDefinitions that are to be 
exported as tiers. If set to NULL (the default) all attribute definitions will be exported as separate tiers.}
}
\description{
Export the annotations of an emuDB to a TextGrid collection (.TextGrid and .wav file pairs)
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB 
# (see ?load_emuDB for more information)

## Export all levels


}

}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

