\name{f2geraus.l}
\alias{f2geraus.l}
\title{A parallel label vector for trackdata f2geraus}
\usage{data(f2geraus.l)}
\description{
   A parallel vector of labels of the language: "aus" for Australian English, "ger" for Standard German.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}