\name{f2geraus}
\alias{f2geraus}
\title{Emu trackdata object}
\usage{data(f2geraus)}
\description{
   Trackdata of F2 of "i:" in Australian English (1 male speaker) and Standard German (1 male speaker). Data taken from the databases andosl06 and kielread06.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}