\name{frames}
\alias{frames}
\title{frames}
\description{
  Get frames from trackdata objects
}
\usage{
frames(trackdata)
}
\arguments{
  \item{trackdata}{an object of class trackdata}
}
\value{
  Data frames from the input object.
}
\seealso{
  \code{\link{trackdata}}
}
\author{Jonathan Harrington}
\keyword{utilities}
