\name{fric.dft}
\alias{fric.dft}
\title{Emu trackdata object}
\usage{data(fric.dft)}
\description{
   Spectral trackdata object (256 point DFT, 16000 Hz sampling freq.)
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}