\name{fric}
\alias{fric}
\title{Emu segment list}
\usage{data(fric)}
\description{
   Segment list of intervocalic [s,z], one speaker, Kiel Corpus of Read speech
}
%\format{
%  NA
%
%}
%


%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}