\name{fric.w}
\alias{fric.w}
\title{Vector of word labels}
\usage{data(fric.w)}
\description{
   Vector of word labels from which the fricatives for segment list fric were taken
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}