% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{get.emuDB.UUID}
\alias{get.emuDB.UUID}
\title{Get UUID of emuDB}
\usage{
get.emuDB.UUID(name = NULL, uuid = NULL)
}
\arguments{
\item{name}{name of emuDB to purge}

\item{uuid}{optional UUID of emuDB}
}
\description{
Returns UUID if emuDB is loaded, throws error otherwise
}
\seealso{
\code{\link{is.emuDB.loaded}}
}

