\name{get.time.element}
\alias{get.time.element}
\title{
Get data for a given time
}
\usage{
get.time.element(timeval, dataset)
}
\description{
Gets data for a given time

}
\arguments{
\item{timeval}{
A time in milliseconds
}
\item{dataset}{
A trackdata object as returned by \code{track}.
}}
\value{
The element number of \code{trackdata$data} corresponding to \code{time} 
}
\seealso{
track, frames
}

\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
