% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_emuDB_UUID}
\alias{get_emuDB_UUID}
\title{Get UUID of emuDB}
\usage{
get_emuDB_UUID(dbName = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Returns UUID if emuDB is loaded, throws error otherwise
}
\seealso{
\code{\link{is.emuDB.loaded}}
}

