% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-crud-helpers.R
\name{get_labelIndex}
\alias{get_labelIndex}
\title{Vectorized function to translate level/attribute name pairs into label indexes.}
\usage{
get_labelIndex(emuDBhandle, levelName, attributeName)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{levelName}{The level of the level/attribute pairs. This vector must
match the \code{attributeName} vector.}

\item{attributeName}{The attribute of the level/attribute pairs. This vector must
match the \code{levelName} vector.}
}
\description{
Vectorized function to translate level/attribute name pairs into label indexes.
}
