% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{get_legalLabels}
\alias{get_legalLabels}
\title{Get legal labels of attributeDefinition of emuDB}
\usage{
get_legalLabels(dbName, levelName, attributeDefinitionName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Get legal labels of attributeDefinition of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

