% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{get_levelCanvasesOrder}
\alias{get_levelCanvasesOrder}
\title{Get levelCanvasesOrder of emuDB}
\usage{
get_levelCanvasesOrder(dbName, perspectiveName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{perspectiveName}{name of perspective}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Get levelCanvasesOrder array that specifies what levels are 
displayed in the according perspective by the EMU-webApp. 
Note that only levels of type 
\code{"SEGMENT"} or \code{"EVENT"} can be displayed as levelCanvases. 
Levels of the type \code{"ITEM"} are always visible in the "show hierrachy"
modal.  For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# get level canvas order of the "default"
# perspective of the "ae" emuDB
get_levelCanvasesOrder(dbName = "ae", 
                       perspectiveName = "default")

}
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

