% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{get_signalCanvasesOrder}
\alias{get_signalCanvasesOrder}
\title{Get signalCanvasesOrder of emuDB}
\usage{
get_signalCanvasesOrder(dbName, perspectiveName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{perspectiveName}{name of perspective}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Get signalCanvasesOrder of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

