% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_UUID}
\alias{get_UUID}
\title{Get UUID of emuDB}
\usage{
get_UUID(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
UUID string
}
\description{
Returns UUID if emuDB is loaded, throws error otherwise. For 
more information on the emuDB format see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

get_UUID(dbName = "ae")

}
}
\seealso{
\code{\link{is.emuDB.loaded}}
}

