% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFiles}
\alias{getFiles}
\title{Expand the base names of a segmentlist so that $utts contains full paths}
\usage{
getFiles(Seglist = NULL, PathToDbRootFolder = NULL, fileExt = NULL)
}
\arguments{
\item{Seglist}{segmentlist to be expandend}

\item{PathToDbRootFolder}{path to root directory (CAUTION: think of DB size and search space!)}

\item{fileExt}{file extention including dot (e.g. '.fms'|'.f0'|'.rms'|...)}
}
\value{
segmentlist with expanded $utts
}
\description{
Recusivly searches a root directory for files matching the
base name specified in the utts files of a segmentlist. If
the utt name is 'XYZ' and fileExt is '.fms' the file
'XYZ.fms' will be searched for in the root directory given.
}
\author{
Raphael Winkelmann
}

