% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{import.mediaFiles.emuDB}
\alias{import.mediaFiles.emuDB}
\title{Import media files to EMU database}
\usage{
import.mediaFiles.emuDB(dbName, dir, targetSessionName = "0000",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dir}{directory containing mediafiles or session directories}

\item{targetSessionName}{name of session in which to create the new bundles}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Import media files to EMU database
}
\examples{
\dontrun{
## Add mediafiles from directory

 import.mediaFiles('myEmuDB',dir="/data/mymedia/")

}
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

