% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.mediaFiles}
\alias{import.mediaFiles}
\title{Import media files to EMU database}
\usage{
import.mediaFiles(dbObj, targetSessionName = "0000", dir)
}
\arguments{
\item{dbObj}{object of class emuDB}

\item{dir}{directory containing mediafiles or session directories}
}
\description{
Import media files to EMU database
}
\examples{
\dontrun{
## Add mediafiles from directory

 import.mediaFiles(emuDB,'0000',dir="/data/mymedia")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{create.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

