% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize_database_tables}
\alias{initialize_database_tables}
\title{Initialize empty database tables}
\usage{
initialize_database_tables(conn, itemsTableName, labelTableName, linksTableName)
}
\arguments{
\item{conn}{sql DB connection}

\item{itemsTableName}{SQL table name of items table}

\item{labelTableName}{SQL table name of label table}
}
\description{
Initialize empty database tables that in the sql database
specified by the conn argument. The tables are items, labels, links
}
\author{
Raphael Winkelmann
}

