% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-crud-helpers.R
\name{insertItemIntoDatabase}
\alias{insertItemIntoDatabase}
\title{Insert one item into the database}
\usage{
insertItemIntoDatabase(emuDBhandle, itemToInsert)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{itemToInsert}{Data frame containing the labels for the item to be inserted.
Must contain the columns \code{session}, \code{bundle}, \code{level},
\code{start_item_seq_idx}, \code{attribute}, \code{labelIndex}, and
\code{label}. The first four of these identify the item and must
contain the same value in all rows. \code{attribute} and \code{labelIndex}
must match up - the label index marks the position of the
attribute within its level (see \code{\link{get_labelIndex}}.}
}
\description{
One item, identified as \code{session:bundle:level:sequenceIndex},
is inserted into the database. One label has to be provided for every attribute
of the given level.
}
