% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{is.emuDB.loaded}
\alias{is.emuDB.loaded}
\title{Test if EMU database is loaded}
\usage{
is.emuDB.loaded(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of EmuDB}
}
\value{
TRUE if loaded, FALSE otherwise
}
\description{
Test if EMU database is loaded
}
\examples{
\dontrun{
## Test if database 'ae' is loaded

  is.emuDB.loaded('ae')
}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

