\name{is.trackdata}
\alias{is.trackdata}
\title{ Test whether an object is an Emu trackdata object }
\description{
  Test whether an object is an Emu trackdata object
}
\usage{
  is.trackdata(data)
}
\arguments{
 \item{data}{ A data object to be tested }
}
\value{
  Returns TRUE if the argument is a trackdata object.
}

\seealso{ \code{\link{emu.track}} }


\keyword{misc}
