% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{isol.l}
\alias{isol.l}
\title{Vector of vowel phoneme labels from the segment list isol}
\format{vector of vowel phoneme labels}
\description{
An EMU dataset
}
\keyword{datasets}

