\name{isol}
\alias{isol}
\docType{data}
\title{ Emu segment list}
\description{
  Segment list of the ausisol06 database of isolated word productions by a male speaker of Australian English.
}
\usage{data(isol)}

}
\details{
  Some simple isolated word data produced by a speaker of Australian English, who produced a number of isolated /dVd/ words where the V is one of the 13 possible monophthongs of Australian English.
}

\seealso{
  \code{\link{segmentlist}}
}


\keyword{emu}
\keyword{datasets}
