\name{keng.dft.5}
\alias{keng.dft.5}
\title{Emu trackdata object}
\usage{data(keng.dft.5)}
\description{
   Spectral matrix parallel to 'keng' obtained with a 512 point DFT at the temporal midpoint of the /k/ aspiration.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}