\name{keng.l}
\alias{keng.l}
\title{Vector of labels}
\usage{data(keng.l)}
\description{
   A vector of labels of the following vowel: "front" when /k/ precedes the above front vowels, "back" when /k/ precedes the above back vowels.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}