\name{keng}
\alias{keng}
\title{Emu segment list}
\usage{data(keng)}
\description{
   Segment list of aspiration of Australian English syllable-initial /k/
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}