\name{label.num}
\alias{label.num}
\title{
Convert labels to integers
}
\usage{
label.num(labs)
}
\arguments{
\item{labs}{
A vector of labels.
}}
\value{
Each label is replaced by a number. 
}
\examples{
 > label.num(c("A", "E", "I", "A, "I", "E" ))
 [1] 1 2 3 1 3 2


}
\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
