% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Perform linear time normalisation on trackdata.}
\usage{
linear(dataset, n = 20)
}
\arguments{
\item{dataset}{A trackdata object as returned by \code{track}.}

\item{n}{The number of points (samples) required for each segment.}
}
\value{
A new trackdata object where the data for each segment has the same
number (\code{n}) of samples.
}
\description{
Performs linear time normalisation on trackdata.
}
\details{
The data for each segment is normalised using the \code{approx} function.
}
\seealso{
approx
}
\keyword{misc}
