% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{list.bundles}
\alias{list.bundles}
\title{List bundles of emuDB}
\usage{
list.bundles(dbName = NULL, session = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{session}{optional session}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object with columns session and name of bundles
}
\description{
List all bundles of emuDB or of particular session.
}

