% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{list.emuDBs}
\alias{list.emuDBs}
\title{List loaded emuDBs}
\usage{
list.emuDBs()
}
\value{
list of emuDBS as data.frame object
}
\description{
Lists overview of loaded emuDBs as data.frame table. Listed columns are name,basePath and UUID
}

