% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.legacy.emu.database.names}
\alias{list.legacy.emu.database.names}
\title{List known database names from legacy EMU installation}
\usage{
list.legacy.emu.database.names()
}
\value{
character vector with database names
}
\description{
List known database names from legacy EMU installation
}
\examples{
\dontrun{
## List legacy EMU known database names

list.legacy.emu.database.names()

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{list.legacy.emu.databases}} \code{\link{convert.legacyEmuDB.by.name.to.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{legacy}

