% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{list.sessions}
\alias{list.sessions}
\title{List sessions of emuDB}
\usage{
list.sessions(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object with session names
}
\description{
List session names of emuDB
}

