% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_attrDefLabelGroups}
\alias{list_attrDefLabelGroups}
\title{List labelGroups of attributeDefinition of emuDB}
\usage{
list_attrDefLabelGroups(dbName, levelName, attributeDefinitionName,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List label groups of attribute definition of emuDB. A label group contains the specifications of a group of 
labels that can be referenced by a name given to the group while querying the emuDB.
For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# list attribute definition label groups
# of attributeDefinition "Phonetic" of the level "Phonetic"
# of the "ae" emuDB
list_attrDefLabelGroups(dbName = "ae", 
                        levelName = "Phonetic" , 
                        attributeDefinitionName = "Phonetic")

}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

