% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_attributeDefinitions}
\alias{list_attributeDefinitions}
\title{List attribute definitions of emuDB}
\usage{
list_attributeDefinitions(dbName, levelName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List attribute definitions of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

