% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_emuDBs}
\alias{list_emuDBs}
\title{List loaded emuDBs}
\usage{
list_emuDBs()
}
\value{
list of emuDBS as data.frame object
}
\description{
Lists overview of loaded emuDBs as data.frame table. Listed columns are name,basePath and UUID
}

