% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.files.R
\name{list_files}
\alias{list_files}
\title{List files of emuDB}
\usage{
list_files(dbName, sessionPattern = NULL, bundlePattern = NULL,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{sessionPattern}{A (glob) pattern matching sessions to be searched from the database}

\item{bundlePattern}{A (glob) pattern matching bundles to be searched from the database}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List files of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

