% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_labelGroups}
\alias{list_labelGroups}
\title{List (global) labelGroups of emuDB}
\usage{
list_labelGroups(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object containing labelGroup infos
}
\description{
List (global) labelGroups of emuDB
}
\author{
Raphael Winkelmann
}

