% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legacy.R
\name{list_legacyEmuDBs}
\alias{list_legacyEmuDBs}
\title{List known databases from legacy EMU installation}
\usage{
list_legacyEmuDBs()
}
\value{
named list with pathes to database template files. The names of the list are the database names
}
\description{
Reads EMU_TEMPLATE_PATH variable from environment or from file ${HOME}/.emu/emu-conf or ${USERPROFILE}/.emu/Emu/emu-conf
and searches for *.tlp template files in this path. The basename of the template file is the database name.
}
\examples{
\dontrun{
## List legacy EMU known databases

list_legacyEmuDBs()

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{list_legacyEmuDBs_names}}
}
\keyword{Emu}
\keyword{database}
\keyword{legacy}

