% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-legacy.R
\name{list_legacyEmuDBs_names}
\alias{list_legacyEmuDBs_names}
\title{List known database names from legacy EMU installation}
\usage{
list_legacyEmuDBs_names()
}
\value{
character vector with database names
}
\description{
List known database names from legacy EMU installation
}
\examples{
\dontrun{
## List legacy EMU known database names

list_legacyEmuDBs_names()

}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{list_legacyEmuDBs}} \code{\link{convert_legacyEmuDB_by_name_to_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{legacy}

