% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_linkDefinitions}
\alias{list_linkDefinitions}
\title{List linkDefinitions of emuDB}
\usage{
list_linkDefinitions(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object containing linkDefinitions infos
}
\description{
List linkDefinitions of emuDB
}
\author{
Raphael Winkelmann
}

