% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{list_perspectives}
\alias{list_perspectives}
\title{List perspectives of emuDB}
\usage{
list_perspectives(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List EMUwebAppConfig$perspectives of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

