% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_ssffTrackDefinitions}
\alias{list_ssffTrackDefinitions}
\title{List ssffTrackDefinitions of emuDB}
\usage{
list_ssffTrackDefinitions(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object containing ssffTrackDefinitions infos
}
\description{
List ssffTrackDefinitions of emuDB
}
\author{
Raphael Winkelmann
}

