% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.database}
\alias{load.database}
\title{Load EMU database}
\usage{
load.database(databaseDir, showProgress = FALSE)
}
\arguments{
\item{databaseDir}{directory of the EMU database}

\item{showProgress}{show progress bar}
}
\value{
object of class emuSX.database
}
\description{
Load EMU database
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuSX}
\keyword{schema}

