% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{load_emuDB}
\alias{load_emuDB}
\title{Load emuDB}
\usage{
load_emuDB(databaseDir, inMemoryCache = TRUE, verbose = TRUE)
}
\arguments{
\item{databaseDir}{directory of the emuDB}

\item{inMemoryCache}{cache the loaded DB in memory}

\item{verbose}{be verbose}
}
\value{
name of emuDB
}
\description{
Load emuDB
}
\examples{
\dontrun{
## Load database 'ae' in directory /homes/mylogin/EMUnew/ae

dbName=load_emuDB(/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

