% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segs.R
\name{make.emuRsegs}
\alias{make.emuRsegs}
\title{Make emuDB segment list}
\usage{
make.emuRsegs(dbName, seglist, query, type)
}
\arguments{
\item{dbName}{name of emuDB}

\item{seglist}{segment list data.frame}

\item{query}{query string}

\item{type}{type of list elements
#}
}
\description{
Make emuDB segment list
}

