% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{modify_levelDefinition}
\alias{modify_levelDefinition}
\title{Modify level definitions of emuDB}
\usage{
modify_levelDefinition(dbName, name, newName = NULL, newType = NULL,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of level definition}

\item{newName}{new name of level definition}

\item{newType}{new type of level definition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Modify level definitions of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

