% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{modify_ssffTrackDefinitions}
\alias{modify_ssffTrackDefinitions}
\title{Modify ssffTrackDefinitions of emuDB}
\usage{
modify_ssffTrackDefinitions(dbName, name, newName, newColumnName = NULL,
  newFileExtension = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{name}{name of ssffTrackDefinitions to be modified}

\item{newName}{new name of ssffTrackDefinitions}

\item{newColumnName}{new columnName of ssffTrackDefinitions}

\item{newFileExtension}{new fileExtension of ssffTrackDefinitions}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Modify ssffTrackDefinitions of emuDB
}

