% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRtrackdata.R
\name{normalize_length.emuRtrackdata}
\alias{normalize_length.emuRtrackdata}
\title{Normalize length of segments contained in an emuRtrackdata object}
\usage{
\method{normalize_length}{emuRtrackdata}(x, N = 21, ...)
}
\arguments{
\item{x}{object of class \code{\link{emuRtrackdata}}}

\item{N}{specify length of normalized segments (each segment in resulting
object will consist of \code{N} rows).}

\item{...}{further arguments (none used by this method)}
}
\value{
emuRtrackdata
}
\description{
Normalize length of segments contained in an emuRtrackdata object
}
\seealso{
\code{\link{emuRtrackdata} \link{emuRsegs}}
}
