% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRtrackdata.R
\name{normalize_length}
\alias{normalize_length}
\title{Normalize length of object}
\usage{
normalize_length(x, ...)
}
\arguments{
\item{x}{object to normalize length for}

\item{...}{further arguments to be passed to the next method.}
}
\value{
object of same class as x which lenght has been normalized
}
\description{
Normalize length of object
}
