\name{percent}
\alias{percent}
\title{
Converts a confusion matrix to a percentage matrix
}
\usage{
percent(data)
}
\arguments{
\item{data}{
A confusion matrix.
}}
\value{
Calculates the percentage of elements, from the total elements in a confusion
matrix row.
}
\seealso{
confusion
}
\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
