% Generated by roxygen2 (4.0.2): do not edit by hand
\name{perform}
\alias{perform}
\title{Performance (hit rate) of a confusion matrix}
\usage{
perform(data)
}
\arguments{
\item{data}{A confusion matrix.}
}
\value{
Caluculates the accuracy (total score) of the confusion matrix,
returning percentage of correct, and incorrect matches.
}
\description{
Performs (hit rate) of a confusion matrix
}
\seealso{
confusion
}
\keyword{misc}

