\name{plos.asp}
\alias{plos.asp}
\title{Vector of times at which the burst onset occurs for segment list plos}
\usage{data(plos.asp)}
\description{
   Vector of times at which the burst onset occurs
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}