\name{plos.lv}
\alias{plos.lv}
\title{Vector of labels of the following vowels for segment list plos}
\usage{data(plos.lv)}
\description{
   A vector of labels of the following vowels of segment list plos
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}