\name{plos.sam}
\alias{plos.sam}
\title{Sampled speech data of plos for segment list plos}
\usage{data(plos.sam)}
\description{
   Sampled speech data of plos
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}