\name{plos.w}
\alias{plos.w}
\title{Vector of word labels for segment list plos}
\usage{data(plos.w)}
\description{
   A vector of the word labels from which the segments were taken
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}