\name{polhom.epg}
\alias{polhom.epg}
\title{EPG-compressed trackdata of the segment list polhom}
\usage{data(polhom.epg)}
\description{
   EPG-compressed trackdata of the segment list polhom (frame rate 10 ms).
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}