\name{polhom.l}
\alias{polhom.l}
\title{Vector of phonetic labels of  the segment list polhom}
\usage{data(polhom.l)}
\description{
   Vector of phonetic labels of  the segment list polhom.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}