\name{polygonplot}
\alias{polygonplot}

\title{polygonplot}
\description{
plots a polygon
}
\usage{
polygonplot(data, labels, order, formant = TRUE, axes = TRUE, xlab = "", ylab = "", main = "", xlim, ylim)
}
\arguments{
  \item{data}{ data matrix}
  \item{labels}{labels}
  \item{order}{ order }
  \item{formant}{ formant TRUE or FALSE transposes the axes}
  \item{axes}{axes}
  \item{xlab}{xlab}
  \item{ylab}{ylab}
  \item{main}{main}
  \item{xlim}{xlim}
  \item{ylim}{ylim}
}
\keyword{ internal }

