% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRtrackdata.R
\name{print.emuRtrackdata}
\alias{print.emuRtrackdata}
\title{Print emuRtrackdata object}
\usage{
\method{print}{emuRtrackdata}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional params}
}
\description{
Print emuRtrackdata object
}
