% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{purge.all.emuDBs}
\alias{purge.all.emuDBs}
\title{Purge emuDB}
\usage{
purge.all.emuDBs(interactive = TRUE)
}
\arguments{
\item{name}{name of emuDB to purge}

\item{uuid}{optional UUID of emuDB}
}
\description{
Purges emuDB from this R session. Does not delete any files of the emuDB.
}

