% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{purge_all_emuDBs}
\alias{purge_all_emuDBs}
\title{Purge all loaded emuDBs}
\usage{
purge_all_emuDBs(interactive = TRUE)
}
\arguments{
\item{interactive}{ask user for confirmation}
}
\description{
Purges emuDB from this R session. Does not delete any files of the emuDB.
}

