% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{purge_emuDB}
\alias{purge_emuDB}
\title{Purge emuDB}
\usage{
purge_emuDB(dbName = NULL, dbUUID = NULL, interactive = TRUE)
}
\arguments{
\item{dbName}{name of emuDB to purge}

\item{dbUUID}{optional UUID of emuDB}

\item{interactive}{ask user for confirmation}
}
\description{
Purges emuDB from this R session. Does not delete any files of the emuDB.
}

