% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query.database}
\alias{query.database}
\title{Query EMU database}
\usage{
query.database(database, query, queryLang = "EQL2", resultType = "emusegs")
}
\arguments{
\item{database}{object of class emuSX.database}

\item{query}{EQL query string}

\item{queryLang}{query language}

\item{resultType}{type (class name) of result}
}
\value{
result set object of class resultType (default EMU seglist 'emusegs')
}
\description{
Query an EMU database
}
\examples{
\dontrun{
## Query database object ae with EQL query [Phonetic=t -> Phonetic=s] and store result seglist in variable segListTs

seglistTs=query.database(ae,"[Phonetic=t -> Phonetic=s]")

## Query seglist from database object ae with EQL query [Syllable=S ^ Phoneme=t]

query.database(ae,"[Syllable=S ^ Phoneme=t]")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.database}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuSX}
\keyword{query}

