% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-server.R
\name{rdaTrackToSsff}
\alias{rdaTrackToSsff}
\title{Converts a signal track from Rda to SSFF}
\usage{
rdaTrackToSsff(rdaPath, columnExpression)
}
\arguments{
\item{rdaPath}{Path to the Rda file containing the signal track}

\item{columnExpression}{The R expression needed to extract the desired signal data from the Rda file, for example `data$audio` or `data$velumgridsigUS`}
}
\value{
Raw contents of an SSFF-encoded file
}
\description{
The SSFF track briefly hits the disk, but the return value of this function is in-memory only.
}
