% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emusegs.R
\name{read.emusegs}
\alias{read.emusegs}
\title{Create an Emu segment list from a file}
\usage{
read.emusegs(file)
}
\arguments{
\item{file}{The name of the file to read}
}
\value{
An Emu segment list.
}
\description{
Create an Emu segment list from a file saved by the Emu query tools.
}
\details{
Reads segment lists created by programs external to R/Splus and stored in
text files on disk.
}
\examples{

## create a segment list file and write it out
seglist.txt <- "database:demo"\\
query:Phonetic=vowel\\
type:segment\\
#\\
}
\author{
Steve Cassidy
}
\seealso{
\code{\link{query}}
}
\keyword{IO}

