% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bundleList.R
\name{read_bundleList}
\alias{read_bundleList}
\title{read bundleList}
\usage{
read_bundleList(emuDBhandle, name)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{name}{name of bundleList (excluding the _bundleList.json suffix)}
}
\value{
tibble with the columns \code{session}, \code{name}, 
\code{comment}, \code{finishedEditing}
}
\description{
read bundleList JSON file in emuDB
}
\details{
Read bundleList JSON file in emuDB that is stored in
the databases root dir sub-dir \code{bundleLists/}
}
