% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{reload}
\alias{reload}
\title{Reload EMU database}
\usage{
reload(db)
}
\arguments{
\item{db}{object of class emuDB}
}
\value{
refreshed database object
}
\description{
Reload an EMU database from disk storage
}
\examples{
\dontrun{
## Reload database object ae

ae=reload(ae)

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

