% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{reload_emuDB}
\alias{reload_emuDB}
\title{Reload EMU database}
\usage{
reload_emuDB(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of EmuDB}
}
\description{
Reload an EMU database from disk storage
}
\examples{
\dontrun{
## Reload database 'ae'

reload_emuDB('ae')
}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

