% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_attrDefLabelGroup}
\alias{remove_attrDefLabelGroup}
\title{Remove labelGroups of attributeDefinition from emuDB}
\usage{
remove_attrDefLabelGroup(dbName, levelName, attributeDefinitionName,
  labelGroupName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{labelGroupName}{name of label group}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Remove labelGroups of attributeDefinition from emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

