% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_attributeDefinition}
\alias{remove_attributeDefinition}
\title{Remove attribute definitions from emuDB}
\usage{
remove_attributeDefinition(dbName, levelName, attributeDefinitionName,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Remove attribute definitions from emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

