% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_labelGroup}
\alias{remove_labelGroup}
\title{Remove (global) labelGroup from emuDB}
\usage{
remove_labelGroup(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of label group}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Remove (global) labelGroup from emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

