% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_levelDefinition}
\alias{remove_levelDefinition}
\title{Remove level definition to emuDB}
\usage{
remove_levelDefinition(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of level definition}

\item{dbUUID}{optional UUID of loaded emuDB}

\item{type}{type of level definition}
}
\description{
Remove level definition to emuDB
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

