% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_linkDefinition}
\alias{remove_linkDefinition}
\title{Remove linkDefinition from emuDB}
\usage{
remove_linkDefinition(dbName, superlevelName, sublevelName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{superlevelName}{name of superlevel of linkDefinition}

\item{sublevelName}{name of sublevel of linkDefinition}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Remove linkDefinition from emuDB
}
\author{
Raphael Winkelmann
}

