% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{remove_perspective}
\alias{remove_perspective}
\title{Remove perspective from emuDB}
\usage{
remove_perspective(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of perspective}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List EMUwebAppConfig$perspective from emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

