% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_ssffTrackDefinition}
\alias{remove_ssffTrackDefinition}
\title{Remove ssffTrackDefinition of emuDB}
\usage{
remove_ssffTrackDefinition(dbName = NULL, name = NULL,
  deleteFiles = FALSE, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{name}{name of ssffTrackDefinitions to be deleted}

\item{deleteFiles}{deletes all files with the fileExtension of the ssffTrackDefinition}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Remove ssffTrackDefinitions of emuDB
}

