% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_ssffTrackDefinitions}
\alias{remove_ssffTrackDefinitions}
\title{Remove ssffTrackDefinitions of emuDB}
\usage{
remove_ssffTrackDefinitions(dbName = NULL, name = NULL,
  deleteFiles = FALSE, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{name}{name of ssffTrackDefinitions to be deleted}

\item{deleteFiles}{deletes all files with the fileExtension of the ssffTrackDefinition}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Remove ssffTrackDefinitions of emuDB
}

