% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-autoproc_annots.R
\name{rename_attributeDefinition}
\alias{rename_attributeDefinition}
\title{Rename a level or attribute definition}
\usage{
rename_attributeDefinition(emuDBhandle, origAttrDef, newAttrDef,
  verbose = FALSE)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{origAttrDef}{name of level/attribute definition in emuDB that is to be changed}

\item{newAttrDef}{new name that shall be assigned to the level/attribute definition}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
While renaming a level or attribute definition, emuR will (1) rewrite the
levelDefinitions in DBconfig, (2) rewrite the linkDefinitions in DBconfig,
(3) rewrite the perspectives in DBconfig, and (4) rewrite all _annot.json
files. (4) May take quite a while, depending on the number of bundles in the
database.
}

