% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-database.R
\name{rename_emuDB}
\alias{rename_emuDB}
\title{Rename emuDB}
\usage{
rename_emuDB(databaseDir, newName)
}
\arguments{
\item{databaseDir}{directory of the emuDB}

\item{newName}{new name of emuDB}
}
\description{
Rename a emuDB. This effectively renames the folder of a 
emuDB the _DBconfig.json file as well as the "name" entry in the _DBconfig.json
file and the _emuDBcache.sqlite file if available.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB
# (see ?load_emuDB for more information)

# rename ae emuDB to "aeNew"
rename_emuDB(databaseDir = "/path/2/ae_emuDB", newName = "aeNew")

}

}
