% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.database.R
\name{requery}
\alias{requery}
\title{Requery EMU database}
\usage{
requery(dbObj, segs, level, targetlevel = level, justlabels = FALSE,
  sequence = 0, longerok = FALSE, resultType = NULL)
}
\arguments{
\item{dbObj}{object of class emuDB}

\item{segs}{segment list}

\item{level}{level of input segment list}

\item{targetlevel}{target level}

\item{justlabels}{labels only}

\item{sequence}{integer which determins sibling position relative to segment in segs}

\item{longerok}{TODO ??}

\item{resultType}{type (class name) of result}
}
\value{
result set object of class resultType (e.g. EMU seglist 'emusegs')
}
\description{
Requery an EMU database
}
\examples{
\dontrun{
## Requery phonetic context of syllable

requery(ae,sl1,level='Syllable',targetlevel='Phonetic')

## Requery previous element on phonetic level

requery(ae,sl2,level='Phonetic',sequence=-1)

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

