% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.database.R
\name{requery_hier}
\alias{requery_hier}
\title{Requery emuR result hierarchical context}
\usage{
requery_hier(seglist, level = NULL, targetLevel = NULL, resultType = NULL,
  dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on}

\item{level}{optional: chearcter string: query to this level first}

\item{targetLevel}{caharcter string: result level}

\item{resultType}{type (class name) of result}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class resultType (default: 'emuRsegs')
}
\description{
Requery an EMU database on particular levels
WARNING! Experimental, syntax and semantics may change!!
}
\examples{
\dontrun{

## Requery phoneme sequence of word (Text) 'beautiful'
sl1=query('ae','Text=beautiful')
requery_hier(sl1,targetLevel='Phoneme')


}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}}
}
\keyword{EQL2}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

