% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.database.R
\name{requery_seq}
\alias{requery_seq}
\title{Requery emuR result context}
\usage{
requery_seq(seglist, offset = 0, offsetRef = "START", length = 1,
  resultType = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on}

\item{offset}{offset in sequence}

\item{offsetRef}{referenec of offset: 'START' start of segments, 'END': end of segments}

\item{length}{item length of returned segment list}

\item{resultType}{type (class name) of result}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class resultType (default: 'emuRsegs')
}
\description{
Requery an EMU database
WARNING! Experimental, syntax and semantics may change!!
}
\examples{
\dontrun{

## Requery previous element of 'p' on phonetic level
sl1=query('ae','Phonetic=p')
contextRequery(sl1,offset=-1)

## Requery context (previuos and following) of 'p' on phonetic level

contextRequery(sl1,offset=-1,length=3)



}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}}
}
\keyword{EQL2}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

