% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-autoproc_annots.R
\name{resample_annots}
\alias{resample_annots}
\title{Resample annotations (\code{_annot.json}) files of emuDB}
\usage{
resample_annots(emuDBhandle, newSampleRate, verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{newSampleRate}{target sample rate}

\item{verbose}{show progress bars and further information}
}
\description{
Resample all annotations (\code{_annot.json}) files of emuDB to a specified 
sample rate. It is up to the user to ensure that the samplerates of 
the annot.json files match those of the \code{.wav} files.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB 
# (see ?load_emuDB for more information)

# resample
resample_annots(ae, newSampleRate = 16000)

}
}
