% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-crud-helpers.R
\name{rewrite_allSequenceIndexes}
\alias{rewrite_allSequenceIndexes}
\title{Rewrite all sequence indexes across all sessions and bundles.}
\usage{
rewrite_allSequenceIndexes(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}
}
\description{
Reads the existing sequences of all items, assuming they are a mixture of
natural values, real values and NULL, sorts them in ascending order, and
replaces them with the sequence 1..n, where n is the number of items on the
respective level in the respective bundle. NULL values are placed at the end
of the sequence.
}
