% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-crud-helpers.R
\name{rewrite_sequenceIndexesOneLevel}
\alias{rewrite_sequenceIndexesOneLevel}
\title{See \code{\link{rewrite_allSequenceIndexes}}}
\usage{
rewrite_sequenceIndexesOneLevel(emuDBhandle, itemsOnLevel)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{itemsOnLevel}{Data frame describing all items on a particular level.}
}
\description{
See \code{\link{rewrite_allSequenceIndexes}}
}
