% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_cunkerFromCano}
\alias{runBASwebservice_cunkerFromCano}
\title{Creates a chunk segmentation using the webservice Chunker.}
\usage{
runBASwebservice_cunkerFromCano(handle, canoLabel, language,
  chunkLabel = "TRN", rootLevel = NULL, chunkLevel = NULL,
  orthoLabel = NULL, chunkerParams = list(), verbose = TRUE,
  perspective = "default", resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{canoLabel}{name of the label (not level!) containing a canonical pronunciation of the words.}

\item{language}{language to be used.}

\item{chunkLabel}{label that will be given to the chunk segmentation. Default: TRN}

\item{rootLevel}{level that will link to the newly created chunk level. Should be the root bundle level.}

\item{orthoLabel}{if there is an orthographic label paired with the canonical pronunciation label, you can provide it here
to get orthographic chunk labels. If NULL, you will end up with chunk labels in SAMPA.}

\item{chunkerParams}{list of parameters to be passed on to the chunker It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Example: g2pParameters = list(minChunkDuration = 10)}

\item{verbose}{Display progress bars and other information}

\item{perspective}{webApp perspective in which the chunk segment tier will be displayed. Default: default.
Set to NULL if you do not want the tier in any perspective.}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to start where the last call failed. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}

\item{orthoLevel}{name of the chunk level that will be created. Defaults to the value of chunkLabel.}
}
\description{
Creates a chunk segmentation using the webservice Chunker.
}

