% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_g2pFromOrtho}
\alias{runBASwebservice_g2pFromOrtho}
\title{Creates SAMPA labels for a tier of tokenized orthographical words.}
\usage{
runBASwebservice_g2pFromOrtho(handle, orthoLabel, language, canoLabel = "KAN",
  g2pParams = list(), verbose = TRUE, resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{orthoLabel}{name of the label (not level!) containing orthographic words.
If this label resides on a segment level, the segment time information is used as a presegmentation.
If it is an item level, no assumption is made about the temporal position of segments.}

\item{language}{language to be used.}

\item{canoLabel}{label that will be given to the canonical form of words. Default: KAN}

\item{g2pParams}{list of parameters to be passed on to G2P. It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling G2P from emuR, and will be internally overridden.
Example: g2pParameters = list(nrm="yes")
Some parameters (such as the output format) cannot be set when calling MAUS from emuR, and will be internally overridden.}

\item{verbose}{Display progress bars and other information}

\item{resume}{If a previous call to run_maus_from_transcription has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to start where the last call failed. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
Creates SAMPA labels for a tier of tokenized orthographical words.
}

