% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_g2pFromTranscription}
\alias{runBASwebservice_g2pFromTranscription}
\title{Creates a word tier with canonical pronunciation in SAMPA from an orthographic transcription.}
\usage{
runBASwebservice_g2pFromTranscription(handle, transcriptionLabel, language,
  canoLabel = "KAN", orthoLabel = "ORT", g2pParams = list(),
  orthoLevel = NULL, verbose = TRUE, resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{transcriptionLabel}{name of the label (not level!) containing an orthographic transcription.
If this label resides on a segment level, the segment time information is used as a chunk segmentation (presegmentation).
If it is an item level, no assumption is made about the temporal position of segments.}

\item{language}{language to be used.}

\item{canoLabel}{label that will be given to the canonical form of words. Default: KAN}

\item{orthoLabel}{label that will be given to the orthographic words. Default: ORT}

\item{g2pParams}{list of parameters to be passed on to G2P. It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling G2P from emuR, and will be internally overridden.
Example: g2pParameters = list(nrm="yes")}

\item{orthoLevel}{name of the word level that will be created. Defaults to the value of orthoLabel.}

\item{verbose}{Display progress bars and other information}

\item{resume}{If a previous call to run_maus_from_transcription has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to start where the last call failed. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
Creates a word tier with canonical pronunciation in SAMPA from an orthographic transcription.
}

