% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_mausFromCano}
\alias{runBASwebservice_mausFromCano}
\title{Creates a phonetic segmentation on the basis of a tokenized word tier with canonical pronunciation in SAMPA, using the segmentation webservice MAUS.}
\usage{
runBASwebservice_mausFromCano(handle, canoLabel, language, mausLabel = "MAU",
  chunkLevel = NULL, mausLevel = NULL, mausParams = NULL,
  verbose = TRUE, perspective = "default", resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{canoLabel}{name of the label (not level!) containing canonical pronunciations of words.
If this label resides on a segment level, the segment time information is used as a presegmentation.
If it is an item level, no assumption is made about the temporal position of segments.}

\item{language}{language to be used.}

\item{mausLabel}{label that will be given to the MAUS segmentation. Default: MAU}

\item{chunkLevel}{if you have a chunk segmentation level, you can provide it to improve the speed and accuracy
of MAUS. The chunk segmentation level must be a segment level, and it must link to the level where orthoLabel resides.}

\item{mausLevel}{name of the level for the MAUS segmentation. Defaults to the value of mausLabel.}

\item{mausParams}{list of parameters to be passed on to MAUS It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Example: mausParameters = list(MODUS="align", RELAXMINDUR="true")
Some parameters (such as the output format) cannot be set when calling MAUS from emuR, and will be internally overridden.}

\item{verbose}{Display progress bars and other information}

\item{perspective}{webApp perspective in which the MAUS segment tier will be displayed. Default: default.
Set to NULL if you do not want the tier in any perspective.}

\item{resume}{If a previous call to run_maus_from_transcription has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to start where the last call failed. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
This function requires a pre-existing tier with words transcribed in SAMPA. It calls the
BAS webservice MAUS to generate a phonemic segmentation. All necessary level, attribute and link definitions
are created in the process. This function requires an internet connection.
}

