% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_mausFromOrtho}
\alias{runBASwebservice_mausFromOrtho}
\title{Creates a phonetic segmentation on the basis of an tokenized orthographic word tier using the segmentation webservices G2P and MAUS.}
\usage{
runBASwebservice_mausFromOrtho(handle, orthoLabel, language,
  canoLabel = "KAN", mausLabel = "MAU", mausLevel = NULL,
  chunkLevel = NULL, mausParams = list(), g2pParams = list(),
  verbose = TRUE, perspective = "default", resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{orthoLabel}{name of the label (not level!) containing orthographic words.
If this label resides on a segment level, the segment time information is used as a presegmentation.
If it is an item level, no assumption is made about the temporal position of segments.}

\item{language}{language to be used. Must be accepted by both the G2P and the MAUS webservices.
If you want to use different language options for G2P and MAUS, please run the services separately
by first calling bas_run_g2p_from_ortho, and then bas_run_maus_from_cano. If you do this, you
must ensure that the label inventories of both languages are compatible.}

\item{canoLabel}{label that will be given to the canonical form of words. Default: KAN}

\item{mausLabel}{label that will be given to the MAUS segmentation. Default: MAU}

\item{mausLevel}{name of the level for the MAUS segmentation. Defaults to the value of mausLabel.}

\item{chunkLevel}{if you have a chunk segmentation level, you can provide it to improve the speed and accuracy
of MAUS. The chunk segmentation level must be a segment level, and it must link to the level where orthoLabel resides.}

\item{mausParams}{list of parameters to be passed on to MAUS It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Example: mausParameters = list(MODUS="align", RELAXMINDUR="true")
Some parameters (such as the output format) cannot be set when calling MAUS from emuR, and will be internally overridden.}

\item{g2pParams}{list of parameters to be passed on to G2P. It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling G2P from emuR, and will be internally overridden.
Example: g2pParameters = list(nrm="yes")}

\item{verbose}{Display progress bars and other information}

\item{perspective}{webApp perspective in which the MAUS segment tier will be displayed. Default: default.
Set to NULL if you do not want the tier in any perspective.}

\item{resume}{If a previous call to run_maus_from_transcription has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to start where the last call failed. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
This function requires a pre-existing tier with tokenized orthographic words. It calls the BAS
webservice G2P to derive a canonical pronunciation from it. On the basis of this canonical pronunciation, it then calls
the BAS webservice MAUS to generate a phonemic segmentation. All necessary level, attribute and link definitions
are created in the process. This function requires an internet connection.
}

