% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_minni}
\alias{runBASwebservice_minni}
\title{Creates a rough phonetic segmentation by running the phoneme decoder webservice MINNI.}
\usage{
runBASwebservice_minni(handle, language, minniLabel = "MINNI",
  rootLevel = NULL, minniLevel = NULL, minniParams = list(),
  verbose = TRUE, perspective = "default", resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{language}{language to be used.}

\item{minniLabel}{label that will be given to the MINNI segmentation. Default: MINNI}

\item{rootLevel}{level that will link to the newly created chunk level. Should be the root bundle level.}

\item{minniParams}{list of parameters to be passed on to MINNI It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling MINNI from emuR, and will be internally overridden.
Example: minniParams = list()}

\item{verbose}{Display progress bars and other information}

\item{perspective}{webApp perspective in which the MINNI segment tier will be displayed. Default: default.
Set to NULL if you do not want the tier in any perspective.}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to retrieve any progress made up to that point. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
Creates a rough phonetic segmentation by running the phoneme decoder webservice MINNI.
}

