% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_pho2sylFromCano}
\alias{runBASwebservice_pho2sylFromCano}
\title{Adds syllabified word labels to a word level that already contains a canonical pronunciation label.}
\usage{
runBASwebservice_pho2sylFromCano(handle, canoLabel, language,
  canoSylLabel = "KAS", pho2sylParams = list(), verbose = TRUE,
  resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{canoLabel}{name of the label (not level!) containing a canonical pronunciation of the words.}

\item{language}{language to be used.}

\item{canoSylLabel}{label that will be given to the new syllabification label. Default: KAS}

\item{pho2sylParams}{list of parameters to be passed on to pho2syl It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling pho2syl from emuR, and will be internally overridden.}

\item{verbose}{Display progress bars and other information}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to retrieve any progress made up to that point. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}
}
\description{
Adds syllabified word labels to a word level that already contains a canonical pronunciation label.
}

