% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_pho2sylFromSegment}
\alias{runBASwebservice_pho2sylFromSegment}
\title{Creates a syllable segmentation on the basis of a phonetic segmentation.}
\usage{
runBASwebservice_pho2sylFromSegment(handle, mausLabel, wordLevel, language,
  sylLabel = "MAS", sylLevel = NULL, pho2sylParams = list(),
  perspective = "default", resume = FALSE, verbose = TRUE)
}
\arguments{
\item{handle}{emuDB handle}

\item{wordLevel}{name of word level. Must be a parent lavel of the segmentation level.}

\item{language}{language to be used.}

\item{sylLabel}{label that will be given to the new syllabification labels. Default: MAS}

\item{sylLevel}{name of the new syllabification level. Defaults to the value of sylLabel.}

\item{pho2sylParams}{list of parameters to be passed on to pho2syl It is your responsibility to ensure that
these parameters are compatible with the webservice API. If they are not, you will likely get an error.
Some parameters (such as the output format) cannot be set when calling pho2syl from emuR, and will be internally overridden.}

\item{perspective}{webApp perspective in which the syllable tier will be displayed. Default: default.
Set to NULL if you do not want the tier in any perspective.}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue),
you can set resume to TRUE for the next call to retrieve any progress made up to that point. This will only work if you have not
run any other emuR functions (such as the query function) in the meantime, as they are likely to delete your temporary data.}

\item{verbose}{Display progress bars and other information}

\item{canoLabel}{name of the label (not level!) containing a phonetic segmentation.}
}
\description{
Creates a syllable segmentation on the basis of a phonetic segmentation.
}

