% Generated by roxygen2 (4.0.2): do not edit by hand
\name{segmentlist}
\alias{emusegs}
\alias{segmentlist}
\title{Segment list}
\format{multi-columned matrix one row per segment
\itemize{
  \item columnlabel
  \item columnsegment onset time
  \item columnsegment offset time
  \item columnutterance name
}}
\description{
A segment list is the result of emu.query() or read.emusegs().
}
\examples{
data(demo.vowels)

   #demo.vowels is a segment list
   demo.vowels
}
\seealso{
\code{\link{emu.query}}, \code{\link{demo.vowels}}
}
\keyword{classes}

