% Generated by roxygen2 (4.0.1): do not edit by hand
\name{serve.emuR.database}
\alias{serve.emuR.database}
\title{Serve EMU database to EMU-Webapp}
\usage{
\method{serve}{emuR.database}(database, port = 8080, debug = FALSE,
  debugLevel = 0)
}
\arguments{
\item{database}{a emuR database object}

\item{port}{the port number to listen on (default: 8080)}

\item{debug}{TRUE to enable debugging (default: no debugging messages)}

\item{debugLevel}{integer higher values generate more detailed debug output}
}
\value{
the possibly modified Emu database object
}
\description{
Starts a websocket server and listens on commands from a running EMU-Webapplication browser GUI instance. The server runs in a loop and blocks the R console. Stop the server with the 'Clear' button of the webapp GUI. The server is also stopped when the browser is closed or the webapplication page is reloaded. Communication is defined by EMU-webApp-websocket-protocol version 0.0.1
}
\examples{
\dontrun{
## Serve database object ae (opens default HTTP/websocket port 8080)

ae=serve(ae)

## serve database object ae, open HTTP/websocket port 9000

ae=serve(ae,port=9000)

}
}
\author{
Klaus Jaensch
}
\keyword{EMU-webapp}
\keyword{Emu}
\keyword{database}
\keyword{emuR}
\keyword{websocket}

