% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/serve_emuDB_files.R
\name{serve_emuDB_files}
\alias{serve_emuDB_files}
\title{Serve files of an EMU database to the EMU-webApp}
\usage{
serve_emuDB_files(path2dbFolder, sessionPattern = ".*",
  bundlePattern = ".*", host = "127.0.0.1", port = 17890, debug = FALSE,
  debugLevel = 0)
}
\arguments{
\item{path2dbFolder}{path to emuDB folder}

\item{sessionPattern}{regex pattern to be used on sessions to be sent to the EMU-webApp}

\item{bundlePattern}{regex pattern to be used on bundles to be sent to the EMU-webApp}

\item{host}{host IP to listen to (default: 127.0.0.1  (localhost))}

\item{port}{the port number to listen on (default: 17890)}

\item{debug}{TRUE to enable debugging (default: no debugging messages)}

\item{debugLevel}{integer higher values generate more detailed debug output}
}
\value{
the possibly modified Emu database object
}
\description{
Naive file server for the EMU-webApp browser GUI \url{http://ips-lmu.github.io/EMU-webApp/}
}
\details{
This function implements a websocket server that implements the EMU-webApp-websocket-protocol in version 0.0.2
(\url{https://github.com/IPS-LMU/EMU-webApp/blob/master/app/manual/EMU-webApp-websocket-protocol.md}).
It is naive in the sense that it strickly serves the apropriate files of a emuDB to the EMU-webApp without
checking their validity or consistency.

Start instructions:
\itemize{
  \item Call this function to start the server.

\item Start a suitable, up-to-date HTML5 capable Web-Browser (Google Chrome, Firefox,...).

\item Navigate to the EMU-webApp URL: \url{http://ips-lmu.github.io/EMU-webApp/}.

\item Press the 'Connect' button in the EMU-webApp and connect to the default URL (unless default values of host or port paramters have been altered).

\item EMU-webApp should load the bundle list and the first bundle of the given database.
}

Stop instructions:
\itemize{
  \item Stop the server with the 'Clear' button of the web application or the reload button of your browser.

\item The server can also be interrupted with Ctrl-C if something goes wrong.
}
}
\examples{
\dontrun{
## serve the files belonging to a emuDB directly to the EMU-webApp (opens default HTTP/websocket port 17890)

serve("/path/to/myDb")
}
}
\author{
Raphael Winkelmann (adapted form original serve function from Klaus Jänsch)
}
\keyword{EMU-webApp}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{websocket}

