% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{set_legalLabels}
\alias{set_legalLabels}
\title{Set legal labels of attributeDefinition of emuDB}
\usage{
set_legalLabels(dbName, levelName, attributeDefinitionName, legalLabels,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{legalLabels}{character array containing legal labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Set legal labels of attributeDefinition of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

