% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{set_levelCanvasesOrder}
\alias{set_levelCanvasesOrder}
\title{Set levelCanvasesOrder of emuDB}
\usage{
set_levelCanvasesOrder(dbName, perspectiveName, order, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{perspectiveName}{name of perspective}

\item{order}{character vector containig names of levelDefinitions}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Set levelCanvasesOrder of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

