% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{set_signalCanvasesOrder}
\alias{set_signalCanvasesOrder}
\title{Set signalCanvasesOrder of emuDB}
\usage{
set_signalCanvasesOrder(dbName, perspectiveName, order, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{perspectiveName}{name of perspective}

\item{order}{character vector containig names of ssffTrackDefinitions}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Set signalCanvasesOrder of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

