\name{sib.l}
\alias{sib.l}
\title{Vector of labels}
\usage{data(sib.l)}
\description{
   A vector of labels of the following vowel: "f" front when [z] precedes "i:", and "I" and "b", back, when [z] precedes "u:", and "U".
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}