\name{sib}
\alias{sib}
\title{Emu segment list}
\usage{data(sib)}
\description{
   Segment list of German syllable-initial [z] preceding front and back vowels, one (male) speaker, Kiel Corpus of Read Speech.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}