\name{sib.w}
\alias{sib.w}
\title{Vector of word labels}
\usage{data(sib.w)}
\description{
   Vector of word labels from which the fricatives for segment list sib were taken
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}