% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRsegs.R
\name{sort.emuRsegs}
\alias{sort.emuRsegs}
\title{Sort emuRsegs segment list by session, bundle and sample_start}
\usage{
\method{sort}{emuRsegs}(x, decreasing, ...)
}
\arguments{
\item{x}{object to print}

\item{decreasing}{NOT IMPLEMENTED!}

\item{...}{additional params}
}
\description{
Sort emuRsegs segment list by session, bundle and sample_start
}
