% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{splitstring}
\alias{splitstring}
\title{Split a string into words.}
\usage{
splitstring(str, char)
}
\arguments{
\item{str}{A string.}

\item{char}{A character to split on}
}
\value{
A vector of strings.  The original \code{str} is split at ever
occurrence of \code{char} to generate a vector of strings.
}
\description{
Splits a string into words.
}
\examples{
splitstring("/home/recog/steve/foo", "/")
#[1] "home" "recog" "steve" "foo"
}
\keyword{misc}

