\name{stops.bark}
\alias{stops.bark}
\docType{data}
\title{Spectral data from the stop release/burst onset to the onset of the following vowel}
\description{
Spectral data calculated with a 512 pt
FFT, frame shift 5 ms from the stop release/burst onset
to the onset of periodicity of the following vowel
with the frequency axis warped to the Bark scale and covering
the frequency interval 200 - 7800 Hz
}
\examples{
data(stops.bark)
## maybe str(stops.bark) ; plot(stops.bark) ...
}
\keyword{datasets}
