\name{stops.dft}
\alias{stops.dft}
\docType{data}
\title{Spectral data from stop release/burst onset
to the onset the following vowel}
\description{
Spectral data calculated with a 512 pt
FFT, frame shift 5 ms from the stop release/burst onset
to the onset of periodicity of the following vowel}
\examples{
data(stops.dft)
}
\keyword{datasets}
