\name{stops10.lab}
\alias{stops10.lab}
\title{A parallel label vector for trackdata stops10}
\usage{data(stops10.lab)}
\description{
   A parallel vector of the stop labels.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}