\name{stops10}
\alias{stops10}
\title{Emu trackdata object}
\usage{data(stops10)}
\description{
   Spectral matrix, 256-point DFT calculated 10 ms after the stop release in word-initial, German [b, d, g] from the gerisol datasbase. Data from one male speaker of standard German. 
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}