% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-database.R
\name{store}
\alias{store}
\title{Store EMU database to directory}
\usage{
store(emuDBhandle, targetDir, options = NULL, verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{targetDir}{target directory}

\item{options}{list of options}

\item{verbose}{show infos and progress bar}

\item{dbUUID}{optional: UUID of database}
}
\description{
Store EMU database to directory
}
\details{
options is a list of key value pairs:
rewriteSSFFTracks if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: FALSE
ignoreMissingSSFFTrackFiles if TRUE missing SSFF track files are ignored, default: FALSE
symbolicLinkSignalFiles if TRUE signal files are symbolic linked instead of copied. Implies: rewriteSSFFTracks=FALSE, Default: FALSE
}
\examples{
\dontrun{
# Store database 'ae' to directory /homes/mylogin/EMUnew/

  store('ae',"/homes/mylogin/EmuStore/")

}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

