% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{summary.emuDB}
\alias{summary.emuDB}
\title{Print summary for EMU database object}
\usage{
summary.emuDB(name = NULL, uuid = NULL)
}
\arguments{
\item{object}{EMU database object}

\item{...}{(not used)}
}
\description{
Gives an overview of an EMU database object
Prints database name, base directory path and informations about annoation levels, attributes, links, and signal file tracks
}

