% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuDBhandle.R
\name{summary.emuDBhandle}
\alias{summary.emuDBhandle}
\title{Print summary of loaded EMU database (emuDB).}
\usage{
\method{summary}{emuDBhandle}(object, ...)
}
\arguments{
\item{object}{emuDBhandle as returned by \code{\link{load_emuDB}}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Gives an overview of an EMU database.
Prints database name, UUID, base directory path, session and bundle 
count and informations about signal track, annotation level, attribute and link definitions.
}
