% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.emusegs}
\alias{summary.emusegs}
\title{summary emusegs}
\usage{
\method{summary}{emusegs}(object, ...)
}
\arguments{
\item{object}{the segmentlist}

\item{\dots}{nothing special}
}
\description{
summarizes data in emu segment lists
}
\keyword{internal}

