% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.trackdata}
\alias{summary.trackdata}
\title{summary trackdata}
\usage{
\method{summary}{trackdata}(object, ...)
}
\arguments{
\item{object}{track data object}

\item{\dots}{see summary}
}
\description{
summarizes trackdata objects
}
\keyword{internal}

