% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.caching.R
\name{update_cache}
\alias{update_cache}
\title{Update cache of emuDB}
\usage{
update_cache(dbName, dbUUID = NULL, verbose = TRUE)
}
\arguments{
\item{dbName}{name of EMU database}

\item{dbUUID}{optional UUID of emuDB}

\item{verbose}{display infos}
}
\description{
Updates sqlite cache of loaded emuDB. This can and should be used
to update changes to precached/loaded DBs as it only updates the deltas 
in the cache which is considerably faster than reloading and therefore 
recacheing the entire DB.
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

