% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{update_ssffTrackDefinitions}
\alias{update_ssffTrackDefinitions}
\title{Update ssffTrackDefinitions of emuDB}
\usage{
update_ssffTrackDefinitions(dbName = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Update ssffTrackDefinitions of emuDB
}

