\name{utt}
\alias{utt}
\title{
Return the utterances portion of a segment list
}
\usage{
utt(segs)
}
\arguments{
\item{segs}{
A mu+ segment list.
}}
\value{
A vector of the utterance names for each segment.
}
\seealso{
mustart, muend, mudur, label
}
\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
