\name{v.sam}
\alias{v.sam}
\title{Emu trackdata object}
\usage{data(v.sam)}
\description{
   512 points of sampled speech data at the temporal midpoint of a German vowel [E]
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}