\name{vowlax.dft.5}
\alias{vowlax.dft.5}
\title{Emu trackdata object}
\usage{data(vowlax.dft.5)}
\description{
Spectral matrix, 16 kHz, DFT window of 256 points
of German lax vowels at the temporal midpoint (parallel to the object 'vowlax')

}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}