% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax.fdat.5}
\alias{vowlax.fdat.5}
\title{Matrix of formant data extracted at the temporal midpoint from the segment
list vowlax.}
\format{matrix of formant data}
\description{
An EMU dataset
}
\keyword{datasets}
