\name{vowlax.fdat.5}
\alias{vowlax.fdat.5}
\title{Formant data extracted from the temporal midpoint of the formant track data vowlax.fdat 
of the segment list vowlax
}
\usage{data(vowlax)}
%\description{
%  weiﬂ nicht
%}
%\format{
%  NA
%
%}
%
%\details{ 
%  NA
%}
%
\seealso{
   \code{\link{segmentlist}}
   \code{\link{trackdata}}   
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}