\name{vowlax.l}
\alias{vowlax.l}
\title{Vector of phoneme labels from segment list vowlax }
\usage{data(vowlax)}
\description{
  For vowlax see data(vowlax)
}
%\format{
%  NA
%
%}
%
\details{ 
   vowlax = emu.query("kielread06", "*", "Kanonic=a | E | I | O")
   vowlax.l = label(vowlax)

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}


\keyword{datasets}
\keyword{emu}