% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax.left}
\alias{vowlax.left}
\title{Vector of labels preceding the vowels from the segment list vowlax}
\format{vector of phoenme labels}
\description{
An EMU dataset
}
\keyword{datasets}

