% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{vowlax.rms.5}
\alias{vowlax.rms.5}
\title{Vector of RMS energy values at the temporal midpoint extracted at the
temporal midpoint from the segment list vowlax}
\format{vector of RMS energy values}
\description{
An EMU dataset
}
\keyword{datasets}

