% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bundleList.R
\name{write_bundleList}
\alias{write_bundleList}
\title{write bundleList}
\usage{
write_bundleList(emuDBhandle, name, bundleList)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{name}{name of bundleList (excluding the _bundleList.json suffix)}

\item{bundleList}{tibble/data.frame with the columns \code{session}, \code{name}, 
\code{comment} (optional), \code{finishedEditing} (optional). Use \link{list_bundles}}
}
\description{
write bundleList JSON file to emuDB
}
\details{
Write bundleList JSON file to emuDB sub-dir \code{bundleLists/}
}
