% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_attrDefLabelGroup}
\alias{add_attrDefLabelGroup}
\title{Add labelGroup of attributeDefinition to emuDB}
\usage{
add_attrDefLabelGroup(dbName, levelName, attributeDefinitionName,
  labelGroupName, labelGroupValues, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{labelGroupName}{name of label group}

\item{labelGroupValues}{character vector of labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add labelGroup of attributeDefinition to emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

