% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_attributeDefinition}
\alias{add_attributeDefinition}
\title{Add attribute definition to emuDB}
\usage{
add_attributeDefinition(dbName, levelName, name, type = "STRING",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{name}{name af new attributeDefinition}

\item{type}{type of new attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add attribute definition to emuDB
}
\author{
Raphael Winkelmann
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

