% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.files.R
\name{add_files}
\alias{add_files}
\title{Add files to emuDB}
\usage{
add_files(dbName, dir, fileExtension, targetSessionName = "0000",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{dir}{directory containing files to be added}

\item{fileExtension}{file extension of file to be added}

\item{targetSessionName}{name of sessions containing 
bundles that the files will be added to}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add files to bundles of specified session of emuDB.
The files that are found in "dir" that have the extension 
"fileExtension" will be copied into the the according bundle
folder that have the same basename as the according file. For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# specify path to folder containing the following
# files we wish to add to: 
# msajc003.zcr, msajc010.zcr, msajc012.zcr, msajc015.zcr, 
# msajc022.zcr, msajc023.zcr and msajc057.zcr 
path2dir = "/path/to/dir/"

# add the files to session "0000" of "ae" emuDB
add_files(dbName = "ae",
          dir = path2dir,
          fileExtension = "zcr",
          targetSessionName = "0000")

}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

