% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_labelGroup}
\alias{add_labelGroup}
\title{Add (global) labelGroup to emuDB}
\usage{
add_labelGroup(dbName, name, values, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of label group}

\item{values}{character vector of labels}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add (global) labelGroup to emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

