% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_levelDefinition}
\alias{add_levelDefinition}
\title{Add level definition to emuDB}
\usage{
add_levelDefinition(dbName, name, type, store = TRUE, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of level definition}

\item{type}{type of level definition}

\item{store}{changes to file system}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add new level definition to emuDB. A level is a more 
general term for what is often referred to as a "tier". 
It is more general in the sense that people usually 
expect tiers to contain time information. Levels 
can either contain time information if they are of the 
type "EVENT" or of the type "SEGMENT" but are timeless 
if they are of the type "ITEM". For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add level called "Phonetic2" to the "ae" emuDB
# that will contain the transcriptions of a second annotator
add_levelDefinition(dbName = "ae", 
                    type = "SEGMENT")

}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

