% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{add_levelDefinition}
\alias{add_levelDefinition}
\title{Add level definition to emuDB}
\usage{
add_levelDefinition(dbName, name, type, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of level definition}

\item{type}{type of level definition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add level definition to emuDB
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

