% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{add_linkDefinition}
\alias{add_linkDefinition}
\title{Add linkDefinition to emuDB}
\usage{
add_linkDefinition(dbName, type, superlevelName, sublevelName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{type}{type of linkDefinition}

\item{superlevelName}{name of super-level of linkDefinition}

\item{sublevelName}{name of sub-level of linkDefinition}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Add new link definition to emuDB. A link definition
specifies the relationship between two levels, the
super-level and the sub-level. The entirety of all link 
definitions of a emuDB specifies the 
hierarchical structure of the database. For more information
on the structural elements of an emuDB see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add link defintition
add_linkDefinition(dbName = "ae",
                   type = "ONE_TO_MANY",
                   superlevelName = "")


}
}

