% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{add_perspective}
\alias{add_perspective}
\title{Add perspective to emuDB}
\usage{
add_perspective(dbName, name, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{name}{name of perspective}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Add perspective to emuDB. The EMU-webApp subdivides different ways 
to look at an emuDB into so called perspectives. These perspectives, 
between which you can switch in the web application, contain 
information on what levels are displayed, which ssffTracks are drawn, 
and so on. For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add perspective called "justTones" to the "ae" emuDB
add_perspective(dbName ="ae",
                name = "justTones") 
                
# add levelCanvasOrder so only the "Tone" level is displayed
set_levelCanvasesOrder(dbName = "ae", 
                       perspectiveName = "justTones", 
                       order = c("Tone"))

}
}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

