% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.legacyEmuDB.by.name.to.emuDB}
\alias{convert.legacyEmuDB.by.name.to.emuDB}
\title{Convert legacy EMU database and store in new format to directory
Loading by name only works if database was used with legacy EMU.}
\usage{
convert.legacyEmuDB.by.name.to.emuDB(dbName, targetDir, verbose = TRUE)
}
\arguments{
\item{dbName}{legacy EMU database name}

\item{targetDir}{target directory}

\item{verbose}{be verbose}
}
\description{
Convert legacy EMU database and store in new format to directory
Loading by name only works if database was used with legacy EMU.
}
\examples{
\dontrun{
## Load database "ae", convert and save in new format to directory /homes/mylogin/EMUnew/

convert.legacyEmuDB.by.name.to.emuDB("ae","/homes/mylogin/EMUnew/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{convert.legacyEmuDB.to.emuDB}} \code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

