% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.legacyEmuDB.to.emuDB}
\alias{convert.legacyEmuDB.to.emuDB}
\title{Convert legacy EMU database and store in new format to directory}
\usage{
convert.legacyEmuDB.to.emuDB(emuTplPath, targetDir, verbose = TRUE)
}
\arguments{
\item{emuTplPath}{EMU template file path}

\item{targetDir}{target directory}

\item{verbose}{be verbose}
}
\description{
Convert legacy EMU database and store in new format to directory
}
\examples{
\dontrun{
## Convert legacy EMU database specified by EMU template file /homes/mylogin/ae/ae.tpl to directory /homes/mylogin/EMUnew/ae

convert.legacyEmuDB.to.emuDB("/homes/mylogin/ae/ae.tpl","/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.database}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuR}
\keyword{schema}

