% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.TextGridCollection.R
\name{convert_TextGridCollection_to_emuDB}
\alias{convert_TextGridCollection_to_emuDB}
\title{Convert a TextGridCollection (.wav & .TextGrid files) to an emuDB}
\usage{
convert_TextGridCollection_to_emuDB(dir, dbName, targetDir,
  tgExt = "TextGrid", audioExt = "wav", tierNames = NULL,
  verbose = TRUE)
}
\arguments{
\item{dir}{path to directory containing TextGridCollection}

\item{dbName}{name given to emuDB}

\item{targetDir}{directory where to save emuDB}

\item{tgExt}{extension of TextGrid files (default=TextGrid meaning filesnames of the form baseName.TextGrid)}

\item{audioExt}{extension of audio files (default=wav meaning filesnames of the form baseName.wav).}

\item{tierNames}{character vector containing names of tiers to extract and convert. If NULL (the default) all
tiers are converted.}

\item{verbose}{display infos & show progress bar}
}
\description{
Converts a TextGridCollection to an emuDB by first generating a file pair list
containing the paths to the .wav & .TextGrid (default extentions) files with the same base
name. It then generates a emuDB DBconfig from the first TextGrid in this file list which specifies
the allowed level names and types in the new emuDB. After this it converts all file pairs to the new format
checking whether they comply to the newly generated DBconfig.
}
\author{
Raphael Winkelmann
}

