% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.filePairList}
\alias{create.filePairList}
\title{Create a file-pair-list}
\usage{
create.filePairList(path2rootDir, ext1, ext2)
}
\arguments{
\item{path2rootDir}{path to root directory (CAUTION: think of DB size and search space!)}

\item{ext1}{first extention to look for}

\item{ext2}{second extention to pair base names of first extention with}
}
\description{
Recursivly searches through a root directory and matches the
basenames of files that have the extentions provided.
}

