% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_emuRdemoData.R
\name{create_emuRdemoData}
\alias{create_emuRdemoData}
\title{Create demo data for emuR package}
\usage{
create_emuRdemoData(dir = tempdir(), precache = F)
}
\arguments{
\item{dir}{directory to create demo data in (default= tempdir())}

\item{precache}{load and purge the create ae database to create an on-file-system cache}
}
\description{
Create a folder in the folder specified
by the dir argument called emuR_demoData.
This folder contains the folders:
\itemize{
\item{ae: }{Containing an emuDB that adheres to the new formant specification (as expected by the \code{\link{load_emuDB}} function)}
\item{TextGrid_collection: }{Containing a TextGrid collection (as expected from the \code{\link{convert_TextGridCollection_to_emuDB}} function)}
\item{legacy_ae: }{Containing legacyEmuDB (as expected by the \code{\link{convert_legacyEmuDB_to_emuDB}} function)}
}
}
\author{
Raphael Winkelmann
}

