% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-crud.R
\name{delete_itemsInLevel}
\alias{delete_itemsInLevel}
\title{Delete new items programmatically}
\usage{
delete_itemsInLevel(emuDBhandle, itemsToDelete, rewriteAllAnnots = TRUE,
  verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{itemsToDelete}{A data frame with the columns
\itemize{
\item\code{session},
\item\code{bundle},
\item\code{level}, and
\item\code{sequenceIndex}.
}
*None* of the columns should be factors.
\code{sequenceIndex} must be numeric (natural-valued), all other columns must
be of type character.}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json
files) (intended for expert use only)}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
Allows to delete annotation items programmatically.
}
