% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{demo.all}
\alias{demo.all}
\title{Emu segment list}
\format{First Column labels Second start time of the segment Third end time
of the segment Fourth utterance name of the utterance the segment was found}
\description{
Segment list of the demo database that is part of the Emu system.  It is
the result of a database query, that searched all segments at level
Phonetic.
}
\details{
A segment list is created via emu.query() or by using the EMU Query Tool.
}
\seealso{
\code{\link{demo.vowels}} \code{\link{segmentlist}}
}
\keyword{datasets}

