% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\name{demo.vowels.f0}
\alias{demo.vowels.f0}
\title{F0 track data for segment list demo.vowels}
\format{An object with \$index, \$ftime and \$data

index: a two columned matrix with the range of the \$data rows that belong
to the segment ftime: a two columned matrix with the times marks of the
segment data: a one columned matrix with the F0 values}
\description{
A track list of the demo database that is part of the Emu system.  It is the
result of get F0 data for the segment list demo.vowels (see
data(demo.vowels)).
}
\details{
A track list is created via the \code{\link{get_trackdata}} function.
}
\seealso{
\code{\link{demo.all.rms}} \code{\link{segmentlist}}
\code{\link{trackdata}}
}
\keyword{datasets}
