% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{dip}
\alias{dip}
\title{Segment list of dipththongs, two speakers one male, one female , Standard
North German, read speech from database kielread}
\format{
segmentlist
}
\description{
An EMU dataset
}
\keyword{datasets}
