% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsmooth.R
\name{dsmooth}
\alias{dsmooth}
\alias{dsmooth.sub}
\title{Smooth the data in a trackdata object.}
\usage{
dsmooth(dataset)
}
\arguments{
\item{dataset}{A trackdata object as returned from \code{track}.}
}
\value{
The result of applying the \code{smooth} function to each column of
the data for each segment in the trackdata object.
}
\description{
Smooths each dataset in a trackdata object using a running mean smoother.
}
\details{
This function uses the \code{dapply} function to apply \code{smooth} to the
data for each segment.
}
\seealso{
smooth, dapply
}
\keyword{misc}

