\name{dur.emusegs}
\alias{dur.emusegs}
\title{ Duration of segments}
\description{
duration of segments is calculated for each segment in the segment list
}
\usage{
\method{dur}{emusegs}(x)
}
\arguments{
  \item{x}{ a segment list}
}
\value{
a vector of durations
}
\author{Jonathan Harrington}
\keyword{ internal }

