% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emusegs.R
\name{dur.emusegs}
\alias{dur.emusegs}
\title{Duration of segments}
\usage{
\method{dur}{emusegs}(x)
}
\arguments{
\item{x}{a segment list}
}
\value{
a vector of durations
}
\description{
duration of segments is calculated for each segment in the segment list
}
\author{
Jonathan Harrington
}
\keyword{internal}

