% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dur.trackdata}
\alias{dur.trackdata}
\title{Duration of trackdata elements}
\usage{
\method{dur}{trackdata}(x)
}
\arguments{
\item{x}{a trackdata object}
}
\value{
a vector of durations
}
\description{
Duration of segments is calculated for each element in the trackdata object
}
\author{
Jonathan Harrington
}
\keyword{internal}

