% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trackdata.class.R
\name{dur.trackdata}
\alias{dur.trackdata}
\title{Duration of trackdata elements}
\usage{
\method{dur}{trackdata}(x)
}
\arguments{
\item{x}{a trackdata object}
}
\value{
a vector of durations
}
\description{
Duration of segments is calculated for each element in the trackdata object
}
\author{
Jonathan Harrington
}
\keyword{internal}

