% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emu.track2}
\alias{emu.track2}
\title{Extract trackdata information for a given segmentlist}
\usage{
emu.track2(Seglist = NULL, FileExtAndTrackName = NULL,
  PathToDbRootFolder = NULL, OnTheFlyFunctionName = NULL,
  OnTheFlyParas = NULL, OnTheFlyOptLogFilePath = NULL)
}
\arguments{
\item{Seglist}{seglist obtained by a function of package seglist with
option newemuutts=T}

\item{FileExtAndTrackName}{file extension and trackname separated by a ':' (e.g. fms:fm where fms is the file extension and fm is the track/column name)}

\item{PathToDbRootFolder}{is the path to the}

\item{OnTheFlyFunctionName}{name of wrassp function to do on-the-fly calculation}

\item{OnTheFlyParas}{a list parameters that will be given to the function
passed in by the OnTheFlyFunctionName parameter. This list can easily be
generated using the \code{formals} function and then setting the according
parameter one wishes to change.}

\item{OnTheFlyOptLogFilePath}{path to log file for on-the-fly function}
}
\value{
an object of type trackdata is returned
}
\description{
A new and improved version of emu.track that utilizes the wrassp package for
signal processing and SSFF file handling.
}
\details{
Reads time relevant data from a given segmentlist, extracts the specified
trackdata and places it into a trackdata object (analogos to the depricated emu.track). The
segmentlist$utts has to either contain valid paths to the signal files in which case the according
SSFF files have to be in the same folder or valid paths to the SSFF files
}
\author{
Raphael Winkelmann
}
\seealso{
\code{\link{formals}}
}
\keyword{misc}

