\name{engassim.l}
\alias{engassim.l}
\title{Vector of phonetic labels for engassim: /nK/ = /nk,ng/, /sK/ = /sk,sg/}
\usage{data(engassim.l)}
\description{
   Vector of phonetic labels for engassim: /nK/ = /nk,ng/, /sK/ = /sk,sg/.
}
%\format{
%  NA
%
%}
%

}
%
\seealso{
   \code{\link{segmentlist}}
}
%
%\examples{
%}

\keyword{emu}
\keyword{datasets}