% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{engassim}
\alias{engassim}
\title{Segment list of a sequence of syllable final n or N preceding k or g ,
isolated words single speaker, Australian English female from database
epgassim.}
\format{segmentlist}
\description{
An EMU dataset
}
\keyword{datasets}

