% Generated by roxygen2 (4.0.1): do not edit by hand
\name{frames}
\alias{frames}
\title{frames}
\usage{
frames(trackdata)
}
\arguments{
\item{trackdata}{an object of class trackdata}
}
\value{
Data frames from the input object.
}
\description{
Get frames from trackdata objects
}
\author{
Jonathan Harrington
}
\seealso{
\code{\link{trackdata}}
}
\keyword{utilities}

