% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdata.class.R
\name{frames}
\alias{frames}
\title{frames}
\usage{
frames(trackdata)
}
\arguments{
\item{trackdata}{an object of class trackdata}
}
\value{
Data frames from the input object.
}
\description{
Get frames from trackdata objects
}
\seealso{
\code{\link{trackdata}}
}
\author{
Jonathan Harrington
}
\keyword{utilities}
