% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{fric}
\alias{fric}
\title{Segment list of word-medial s or z one male speaker of Standard North
German, read speech from database kielread.}
\format{
segmentlist
}
\description{
An EMU dataset
}
\keyword{datasets}
