% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfuns.R
\name{get.time.element}
\alias{get.time.element}
\title{Get data for a given time}
\usage{
get.time.element(timeval, dataset)
}
\arguments{
\item{timeval}{A time in milliseconds}

\item{dataset}{A trackdata object as returned by \code{track}.}
}
\value{
The element number of \code{trackdata$data} corresponding to
\code{time}
}
\description{
Gets data for a given time
}
\seealso{
track, frames
}
\keyword{misc}

