% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{get_legalLabels}
\alias{get_legalLabels}
\title{Get legal labels of attributeDefinition of emuDB}
\usage{
get_legalLabels(dbName, levelName, attributeDefinitionName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Get legal labels of a specific attributeDefinition of a emuDB. The legal labels are an array 
of strings that specifies the labels that are legal (i.e. allowed / valid) for the given attribute. 
As the EMU-webApp won't allow the annotator to enter any labels that are not 
specified in this array, this is a simple way of assuring that a level 
has a consistent label set. For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# get legal labels of the 
# default "Phonetic" attributeDefinition of
# the "Phonetic" level of "ae" emuDB
get_legalLabels(dbName = "ae", 
                levelName = "Phonetic", 
                attributeDefinitionName = "Phonetic")
                
# NOTE: this will return NA as there are
# no legalLabels defined for this or any other
# attributeDefinition of the "ae" emuDB

}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

