% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-get_trackdata.R
\name{get_trackdata}
\alias{emu.track}
\alias{get_trackdata}
\title{Get trackdata from loaded emuDB}
\usage{
get_trackdata(emuDBhandle, seglist = NULL, ssffTrackName = NULL,
  cut = NULL, npoints = NULL, onTheFlyFunctionName = NULL,
  onTheFlyParams = NULL, onTheFlyOptLogFilePath = NULL,
  nrOfAllocationRows = 10000, resultType = "trackdata", verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{seglist}{\code{\link{emuRsegs}} or \code{\link{emusegs}} object obtained by \code{\link{query}}ing a loaded emuDB}

\item{ssffTrackName}{The name of track that one wishes to extract (see 
\code{\link{list_ssffTrackDefinitions}} for the defined ssffTracks of the 
emuDB). If the parameter \code{onTheFlyFunctionName} is set, then 
this corresponds to the column name af the AsspDataObj (see
\code{wrasspOutputInfos[[onTheFlyFunctionName]]$tracks} and \code{\link{wrasspOutputInfos}}). 
If the parameter \code{onTheFlyFunctionName} is set and this one isn't, then per default
the first track listed in the \code{wrasspOutputInfos} is chosen (\code{wrasspOutputInfos[[onTheFlyFunctionName]]$tracks[1]}).}

\item{cut}{An optional cut time for segment data, ranges between 
0 and 1, a value of 0.5 will extract data only at the segment midpoint.}

\item{npoints}{An optional number of points to retrieve for each segment or event. 
For segments this requires the \code{cut} parameter to be set; if this is the case, then data is extracted around the resulting cut time. 
For events data is extracted around the event time. If npoints is an odd number, the 
samples are centered around the cut-time-sample; if not, they are skewed to the
right by one sample.}

\item{onTheFlyFunctionName}{Name of wrassp function that will perform the on-the-fly 
calculation (see \code{names(wrasspOutputInfos)} for a list of all the signal processing functions wrassp provides)}

\item{onTheFlyParams}{A \code{pairlist} of parameters that will be given to the function 
passed in by the \code{onTheFlyFunctionName} parameter. This list can easily be 
generated by applying the \code{formals} function to the on-the-fly function name and then setting the according 
parameter one wishes to change.}

\item{onTheFlyOptLogFilePath}{Path to optional log file for on-the-fly function}

\item{nrOfAllocationRows}{If this size limit of the data matrix is reached 
a further \code{nrOfAllocationRows} more rows will be allocated. As this allocation leads to
a performance penalty one should consider increasing this number for large emuDBs.}

\item{resultType}{Specify class of returned object. Either \code{"emuRtrackdata"} or \code{"trackdata"}.}

\item{verbose}{Show progress bars and further information}
}
\value{
If the \code{cut} parameter is not set (the default) an object of type \code{\link{trackdata}} or \code{\link{emuRtrackdata}} 
is returned. If \code{cut} is set and \code{npoints} is not, or the seglist 
is of type event and npoints is not set, a \code{\link{data.frame}} is returned.
}
\description{
Extract trackdata information from a loaded emuDB that 
corresponds to the entries in a segment list.
}
\details{
This function utilizes the wrassp package for signal processing and 
SSFF/audio file handling. It reads time relevant data from a given 
segment list (\code{\link{emuRsegs}} or \code{\link{emusegs}}), extracts the 
specified trackdata and places it into a 
trackdata object (analogous to the depricated \code{emu.track}). This function
replaces the deprecated \code{emu.track} function. For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB 
# (see ?load_emuDB for more information)

# query loaded "ae" emuDB for all "i:" segments of the "Phonetic" level
sl = query(emuDBhandle = ae, 
           query = "Phonetic == i:")

# get the corresponding formant trackdata
td = get_trackdata(emuDBhandle = ae, 
                   seglist = sl, 
                   ssffTrackName = "fm")

# get the corresponding F0 trackdata
# as there is no F0 ssffTrack defined in the "ae" emuDB we will 
# calculate the necessary values on-the-fly
td = get_trackdata(emuDBhandle = ae, 
                   seglist = sl, 
                   onTheFlyFunctionName = "ksvF0")
                   
}

}
\seealso{
\code{\link{formals}}, \code{\link{wrasspOutputInfos}}, \code{\link{trackdata}}, \code{\link{emuRtrackdata}}
}
\keyword{misc}

