% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{import.mediaFiles}
\alias{import.mediaFiles}
\title{Import media files to EMU database}
\usage{
import.mediaFiles(db, dir, targetSessionName = "0000")
}
\arguments{
\item{db}{object of class emuDB}

\item{dir}{directory containing mediafiles or session directories}

\item{targetSessionName}{name of session in which to create the new bundles}
}
\value{
modified database object
}
\description{
Import media files to EMU database
}
\examples{
\dontrun{
## Add mediafiles from directory

 import.mediaFiles(emuDB,'0000',dir="/data/mymedia")

}
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

