% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sqlDBactions.R
\name{initialize_database_tables}
\alias{initialize_database_tables}
\title{Initialize empty database tables}
\usage{
initialize_database_tables(conn, itemsTableName, labelTableName, linksTableName)
}
\arguments{
\item{conn}{sql DB connection}

\item{itemsTableName}{SQL table name of items table}

\item{labelTableName}{SQL table name of label table}

\item{linksTableName}{SQL table name of links table}
}
\description{
Initialize empty database tables in the sql database
specified by the conn argument. The tables are items, labels, links
}
\author{
Raphael Winkelmann
}

