% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralclass.R
\name{is.spectral}
\alias{is.spectral}
\title{Function to test whether the object is of class "spectral"}
\usage{
is.spectral(dat)
}
\arguments{
\item{dat}{An R object}
}
\value{
A single element logical vector: TRUE or FALSE
}
\description{
Returns TRUE or FALSE depending on whether the object is of class "spectral"
}
\examples{


is.spectral(vowlax.dft.5)
is.spectral(fric.dft)
is.spectral(fric.dft$data)
is.spectral(vowlax.dft.5[1,])
is.spectral(fric.dft[1,1])



}
\seealso{
\code{\link{as.spectral}}
}
\author{
Jonathan Harrington
}
\keyword{attribute}
