% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.trackdata}
\alias{is.trackdata}
\title{Test whether an object is an Emu trackdata object}
\usage{
is.trackdata(object)
}
\arguments{
\item{object}{A data object to be tested}
}
\value{
Returns TRUE if the argument is a trackdata object.
}
\description{
Test whether an object is an Emu trackdata object
}
\seealso{
\code{\link{emu.track}}
}
\keyword{misc}

