% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{isol}
\alias{isol}
\title{Segment list of vowels in a d d context isolated word speech, one male
speaker of Australian English from database isolated.}
\format{segmentlist}
\description{
An EMU dataset
}
\keyword{datasets}

