% Generated by roxygen2 (4.0.1): do not edit by hand
\name{label}
\alias{label}
\alias{label.emusegs}
\alias{utt}
\alias{utt.emusegs}
\title{Get labels / utterances from segment list}
\usage{
label(segs)
}
\arguments{
\item{segs}{segment list}
}
\value{
label / utterance vector
}
\description{
label: extracts the labels from the segment list.  utt: extracts the
utterances from the segment list.
}
\examples{
data(dip)
   #dip is a segment list - first ten segments only
   dip[1:10,]


   #extract labels from the segment list
   dips.labs = label(dip)
   dips.labs
}
\author{
Jonathan Harrington
}
\seealso{
\code{\link{segmentlist} \link{start} \link{end}}
}
\keyword{methods}

