% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_attrDefLabelGroups}
\alias{list_attrDefLabelGroups}
\title{List labelGroups of attributeDefinition of emuDB}
\usage{
list_attrDefLabelGroups(dbName, levelName, attributeDefinitionName,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List labelGroups of attributeDefinition of emuDB
}
\author{
Raphael Winkelmann
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

