% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_attributeDefinitions}
\alias{list_attributeDefinitions}
\title{List attribute definitions of emuDB}
\usage{
list_attributeDefinitions(dbName, levelName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List attribute definitions of emuDB. Currently the only type of attribute
definition that is supported is a label (`"type": "STRING"`). As it is possible to define
multiple attribute definitions per level it is possible to
have multiple parallel labels in a single level. This means that a single annotation 
item instance can contain multiple labels while sharing other 
properties such as the start and duration information. Note that there is always one
default attribute definitions present that has the same name as the level it belongs 
to. For more information on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# list attribute definitions for level "Word"
# of the "ae" emuDB
list_attributeDefinitions(dbName = "ae", 
                          levelName = "Word")
                          
}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

