% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{list_bundleFilePaths}
\alias{list_bundleFilePaths}
\title{List file paths emuDBs bundles}
\usage{
list_bundleFilePaths(dbName, fileExtention, sessionPattern = ".*",
  bundlePattern = "*", dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{fileExtention}{file extention of files}

\item{sessionPattern}{A (regex) pattern matching sessions of emuDB}

\item{bundlePattern}{A (regex) pattern matching bundles of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
list of emuDBS as data.frame object
}
\description{
Lists file paths of files belonging to emuDB
}

