% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-database.R
\name{list_bundleFilePaths}
\alias{list_bundleFilePaths}
\title{List file paths of emuDBs bundles}
\usage{
list_bundleFilePaths(emuDBhandle, fileExtention, sessionPattern = ".*",
  bundlePattern = "*")
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{fileExtention}{file extention of files}

\item{sessionPattern}{A (regex) pattern matching sessions of emuDB}

\item{bundlePattern}{A (regex) pattern matching bundles of emuDB}
}
\value{
file paths as character vector
}
\description{
List file paths of files belonging to emuDB.  For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB 
# (see ?load_emuDB for more information)

# list all .fms file paths of ae emuDB
list_bundleFilePaths(emuDBhandle = ae, 
                     fileExtention = "fms") 

}

}

