% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_bundles}
\alias{list_bundles}
\title{List bundles of emuDB}
\usage{
list_bundles(dbName, session = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{session}{optional session}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object with columns session and name of bundles
}
\description{
List all bundles of emuDB or of particular session.
}

