% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-database.R
\name{list_bundles}
\alias{list_bundles}
\title{List bundles of emuDB}
\usage{
list_bundles(emuDBhandle, session = NULL)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{session}{optional session}
}
\value{
data.frame object with columns session and name of bundles
}
\description{
List all bundles of emuDB or of particular session.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB
# (see ?load_emuDB for more information)

# list bundles of session "0000" of ae emuDB
list_bundles(emuDBhandle = ae,
             session = "0000")

}

}

