% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.files.R
\name{list_files}
\alias{list_files}
\title{List files of emuDB}
\usage{
list_files(dbName, sessionPattern = "*", bundlePattern = "*",
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{sessionPattern}{A (glob) pattern matching sessions to be searched from the database}

\item{bundlePattern}{A (glob) pattern matching bundles to be searched from the database}

\item{dbUUID}{optional UUID of loaded emuDB, in case dbName is ambiguous}
}
\value{
file paths as character vector
}
\description{
List files belonging to emuDB. For 
more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB
# (see ?load_emuDB for more information)

# list all files of "ae" emuDB
list_files(dbName = "ae") 

# list all files of "ae" emuDB in bundles ending with '3'
list_files(dbName = "ae",bundlePattern="*3") 

}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

