% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_levelDefinitions}
\alias{list_levelDefinitions}
\title{List level definitions of emuDB}
\usage{
list_levelDefinitions(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\value{
\code{\link{data.frame}} containing name, type and nrOfAttrDefs of each level definition
}
\description{
List level definitions of emuDB . A level is an essential structural element of any 
emuDB. A level is a more general term for what is often referred to as a "tier". It 
is more general in the sense that people usually expect tiers to contain time 
information. Levels can either contain time information if they are of the 
type "EVENT" or of the type "SEGMENT" but are timeless if they are of the type "ITEM". 
For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# list level definition of "ae" emuDB
list_levelDefinitions(dbName = "ae")

}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

