% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{list_levelDefinitions}
\alias{list_levelDefinitions}
\title{List level definitions of emuDB}
\usage{
list_levelDefinitions(dbName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
List level definitions of emuDB
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

