% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_sessions}
\alias{list_sessions}
\title{List sessions of emuDB}
\usage{
list_sessions(dbName = NULL, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
data.frame object with session names
}
\description{
List session names of emuDB
}

