% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-database.R
\name{list_sessions}
\alias{list_sessions}
\title{List sessions of emuDB}
\usage{
list_sessions(emuDBhandle)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}
}
\value{
data.frame object with session names
}
\description{
List session names of emuDB
}
\examples{
\dontrun{

##################################
# prerequisite: loaded ae emuDB
# (see ?load_emuDB for more information)

# list all sessions of ae emuDB
list_sessions(emuDBhandle = ae)

}

}

