% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.emuDB}
\alias{load.emuDB}
\title{Load EMU database}
\usage{
load.emuDB(databaseDir, verbose = TRUE)
}
\arguments{
\item{databaseDir}{directory of the EMU database}

\item{verbose}{be verbose}
}
\value{
object of class emuR.database
}
\description{
Load EMU database
}
\examples{
\dontrun{
## Load database 'ae' in directory /homes/mylogin/EMUnew/ae

ae=load.emuDB("ae","/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuR}
\keyword{schema}

