% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_emuDB}
\alias{load_emuDB}
\title{Load emuDB}
\usage{
load_emuDB(databaseDir, inMemoryCache = FALSE, verbose = TRUE)
}
\arguments{
\item{databaseDir}{directory of the emuDB}

\item{inMemoryCache}{cache the loaded DB in memory}

\item{verbose}{be verbose}
}
\value{
name of emuDB
}
\description{
Load emuDB
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

