% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{load_emuDB}
\alias{load_emuDB}
\title{Load EMU database}
\usage{
load_emuDB(databaseDir, verbose = TRUE)
}
\arguments{
\item{databaseDir}{directory of the EMU database}

\item{verbose}{be verbose}
}
\value{
name of emuDB
}
\description{
Load EMU database
}
\examples{
\dontrun{
## Load database 'ae' in directory /homes/mylogin/EMUnew/ae

ae=load_emuDB("ae","/homes/mylogin/EMUnew/ae")

}
}
\author{
Klaus Jaensch
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

