% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRsegs.R
\name{make.emuRsegs}
\alias{make.emuRsegs}
\title{Make emuDB segment list}
\usage{
make.emuRsegs(dbName, seglist, query, type)
}
\arguments{
\item{dbName}{name of emuDB}

\item{seglist}{segment list data.frame}

\item{query}{query string}

\item{type}{type of list elements}
}
\description{
Make emuDB segment list
}
