% Generated by roxygen2 (4.0.1): do not edit by hand
\name{muclass}
\alias{muclass}
\title{Find common elements in vectors}
\usage{
muclass(labels, class)
}
\arguments{
\item{labels}{A vector of labels.}

\item{class}{A label or vector of labels.}
}
\value{
A logical vector which is T for each element in \code{labels} which
matches \code{class} or an element of \code{class}.
}
\description{
Finds common elements in vectors
}
\examples{
muclass(c("a", "b", "c"), c("a", "c"))
#[1] T F T
}
\seealso{
match
}
\keyword{misc}

