% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRtrackdata.R
\name{normalize_length}
\alias{normalize_length}
\title{Normalize length of segments contained in a \code{data.frame} like object returned by \code{\link{get_trackdata}}}
\usage{
normalize_length(x, colNames = NULL, N = 21)
}
\arguments{
\item{x}{data.frame like object that was generated by \code{\link{get_trackdata}} with 
the resultType set to either \code{emuRtrackdata} or \code{tibble}}

\item{colNames}{character vector containing names of columns to normalize. If not set all 
data columns are normalized (T1-TN as well as other numeric columns).}

\item{N}{specify length of normalized segments (each segment in resulting
object will consist of \code{N} rows).}
}
\value{
data.frame like object containing the length normalized segments
}
\description{
Normalize length of segments contained in a \code{data.frame} like object returned by \code{\link{get_trackdata}}
}
\seealso{
\code{\link{emuRtrackdata} \link{emuRsegs}}
}
