\name{perform}
\alias{perform}
\title{
Performance (hit rate) of a confusion matrix
}
\usage{
perform(data)
}
\arguments{
\item{data}{
A confusion matrix.
}}
\value{
Caluculates the accuracy (total score) of the confusion matrix, returning
percentage of correct, and incorrect matches.
}
\seealso{
confusion
}
\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
