% Generated by roxygen2 (4.0.1): do not edit by hand
\name{polygonplot}
\alias{polygonplot}
\title{polygonplot}
\usage{
polygonplot(data, labels, order, formant = TRUE, axes = TRUE, xlab = "",
  ylab = "", main = "", xlim, ylim)
}
\arguments{
\item{data}{data matrix}

\item{labels}{labels}

\item{order}{order}

\item{formant}{formant TRUE or FALSE transposes the axes}

\item{axes}{axes}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{main}{main}

\item{xlim}{xlim}

\item{ylim}{ylim}
}
\description{
plots a polygon
}
\keyword{internal}

