% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuRsegs.R
\name{print.emuRsegs}
\alias{print.emuRsegs}
\title{Print emuDB segment list}
\usage{
\method{print}{emuRsegs}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional params}
}
\description{
Print emuDB segment list
}
