% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{purge.emuDB}
\alias{purge.emuDB}
\title{Purge emuDB}
\usage{
purge.emuDB(dbName = NULL, dbUUID = NULL, interactive = TRUE)
}
\arguments{
\item{dbName}{name of emuDB to purge}

\item{dbUUID}{optional UUID of emuDB}

\item{interactive}{}
}
\description{
Purges emuDB from this R session. Does not delete any files of the emuDB.
}

