% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.database.R
\name{query}
\alias{query}
\title{Query EMU database}
\usage{
query(dbName = NULL, query, sessionPattern = NULL, bundlePattern = NULL,
  queryLang = "EQL2", timeRefSegmentLevel = NULL, resultType = NULL,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of EMU database}

\item{query}{EQL2 query string}

\item{sessionPattern}{A regular expression pattern matching session names to be searched from the database}

\item{bundlePattern}{A regular expression pattern matching bundle names to be searched from the database}

\item{queryLang}{query language}

\item{timeRefSegmentLevel}{set time derivation segment level}

\item{resultType}{type (class name) of result}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
result set object of class resultType (default: \link{emuRsegs}, compatible to legacy type \link{emusegs})
}
\description{
Query an EMU database
}
\details{
Evaluates EQL2 query string on emuDB and returns a segment list of desired type.  
For details of the query language please refer to EQL vignette (type: \code{vignette('EQL')} ). 
Time information of symbolic elements (type 'ITEM') are derived from linked segment levels if available. If not available time and sample values may be set to \code{\link{NA}}.
}
\examples{
\dontrun{
## Query database 'ae' using EQL query [Phonetic=t -> Phonetic=s] 
## and store result seglist in variable segListTs

seglistTs=query('ae','[Phonetic=t -> Phonetic=s]')

## Query seglist from database ae using EQL query [Syllable=S ^ Phoneme=t]
## Request legacy result type 'emusegs'

query('ae','[Syllable=S ^ Phoneme=t]',resultType='emusegs')


## Query 'p' phonemes from bundles whose bundle names start with 'msajc07' 
## and whose session names start with '00'

query('ae','Phoneme=p',bundlePattern='msajc05.*',sessionPattern ='00.*')


}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

