% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.database.R
\name{query}
\alias{query}
\title{Query emuDB}
\usage{
query(dbName = NULL, query, sessionPattern = NULL, bundlePattern = NULL,
  queryLang = "EQL2", timeRefSegmentLevel = NULL, resultType = NULL,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of EMU database}

\item{query}{EQL2 query string}

\item{sessionPattern}{A regular expression pattern matching session names to be searched from the database}

\item{bundlePattern}{A regular expression pattern matching bundle names to be searched from the database}

\item{queryLang}{query language}

\item{timeRefSegmentLevel}{set time derivation segment level}

\item{resultType}{type (class name) of result}

\item{dbUUID}{optional UUID of emuDB}
}
\value{
result set object of class resultType (default: \link{emuRsegs}, compatible to legacy type \link{emusegs})
}
\description{
Function to query an emuDB
}
\details{
Evaluates EQL2 query string on an emuDB referenced by dbName and returns a segment list of desired type.  
For details of the query language please refer to EQL vignette (type: \code{vignette('EQL')} ).
Returns a list of segments which meet the conditions given by the query string. A segment is determined by a start and end item. Segments with item length one have the same start and end item.
Segment objects contain the start and end time information of the segment. \link{emuRsegs} objects additionally contain sample position of start and end item. 
Time information of symbolic elements (type 'ITEM') are derived from linked segment levels if available. If time and sample values cannot be derived they will be set to \code{\link{NA}}.
\link{emuRsegs} result lists will be ordered by the hidden columns UUID,session,bundle and sample start position. Legacy \link{emusegs} lists are ordered by the columns utts and start.
The query may be limited to session and/or bundle names specified by regular expression pattern strings (see \link{regex}) for parameters \code{sessionPattern} respectively \code{bundlePattern}.
}
\examples{
\dontrun{
## Query database 'ae' using EQL query [Phonetic=t -> Phonetic=s] 
## and store result seglist in variable segListTs

seglistTs=query('ae','[Phonetic=t -> Phonetic=s]')

## Query seglist from database ae using EQL query [Syllable=S ^ Phoneme=t]
## Request legacy result type 'emusegs'

query('ae','[Syllable=S ^ Phoneme=t]',resultType='emusegs')


## Query 'p' phonemes from bundles whose bundle names start with 'msajc07' 
## and whose session names start with '00'

query('ae','Phoneme=p',bundlePattern='msajc05.*',sessionPattern ='00.*')


}

}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

