% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query}
\alias{query}
\title{Query EMU database}
\usage{
query(dbObj, query, queryLang = NULL, resultType = NULL)
}
\arguments{
\item{dbObj}{object of class emuDB}

\item{query}{EQL query string}

\item{queryLang}{query language}

\item{resultType}{type (class name) of result}
}
\value{
result set object of class resultType (e.g. EMU seglist 'emusegs')
}
\description{
Query an EMU database
}
\examples{
\dontrun{
## Query database object ae with EQL query [Phonetic=t -> Phonetic=s] and store result seglist in variable segListTs

seglistTs=query(ae,"[Phonetic=t -> Phonetic=s]",resultType='emusegs')

## Query seglist from database object ae with EQL query [Syllable=S ^ Phoneme=t]

query(ae,"[Syllable=S ^ Phoneme=t]",resultType='emusegs')

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{EQL}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{query}

