\name{radians}
\alias{radians}
\title{
Converts degrees to radians
}
\usage{
radians(degrees)
}
\description{}
\arguments{
\item{degrees}{
Angular measurement for conversion.
}}
\value{
Angular measurement in radians.
}
\details{
There are 360 degrees or 2 * PI radians in one full rotation.
}
\keyword{misc}
% Converted by Sd2Rd version 0.3-3.
