% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{radians}
\alias{radians}
\title{Converts degrees to radians}
\usage{
radians(degrees)
}
\arguments{
\item{degrees}{Angular measurement for conversion.}
}
\value{
Angular measurement in radians.
}
\description{
Converts degrees to radians
}
\details{
There are 360 degrees or 2 * PI radians in one full rotation.
}
\keyword{misc}

