% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_attrDefLabelGroup}
\alias{remove_attrDefLabelGroup}
\title{Remove labelGroup of attributeDefinition of emuDB}
\usage{
remove_attrDefLabelGroup(dbName, levelName, attributeDefinitionName,
  labelGroupName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{labelGroupName}{name of label group}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Remove label group that was previously added by 
\code{\link{add__attrDefLabelGroup}}. For more information 
on the structural elements of an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

sampaNasals = c("m", "F", "n", "J", "N")

# add these values to the default Phonetic attribute
# definition of the "Phonetic" level of the "ae" emuDB
add_attrDefLabelGroup(dbName = "ae",
                      levelName = "Phonetic",
                      attributeDefinitionName = "Phonetic",
                      labelGroupName = "sampaNasals",
                      labelGroupValues = sampaNasals)
                      
# remove the newly added attrDefLabelGroup
remove_attrDefLabelGroup(dbName = "ae",
                         levelName = "Phonetic",
                         attributeDefinitionName = "Phonetic",
                         labelGroupName = "sampaNasals")

}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

