% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.R
\name{remove_attributeDefinition}
\alias{remove_attributeDefinition}
\title{Remove attribute definitions from emuDB}
\usage{
remove_attributeDefinition(dbName, levelName, attributeDefinitionName,
  dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{levelName}{name of level}

\item{attributeDefinitionName}{name of attributeDefinition}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Remove additional attribute definition from an existing level definition
of a emuDB that was added with the \code{\link{add_attributeDefinition}} 
function. For more information on the structural elements of 
an emuDB see \code{vignette(emuDB)}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# add additional attribute definition to the "Phonetic" level
# of the "ae" emuDB that will contain the UTF8 IPA
# symbols of the phonetic transcriptions
add_attributeDefinition(dbName = "ae",
                        levelName = "Phonetic",
                        name = "IPA-UTF8")

# remove newly added attributeDefinition
remove_attributeDefinition(dbName = "ae",
                           levelName = "Phonetic",
                           name = "IPA-UTF8")
}

}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{schema}

