% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requery.database.R
\name{requery_hier}
\alias{requery_hier}
\title{Requery hierarchical context of segment list}
\usage{
requery_hier(seglist, level = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on (type: 'emuRsegs')}

\item{level}{character string: result level}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class 'emuRsegs'
}
\description{
Requery hierarchical context of segment list
}
\examples{
\dontrun{

## Downward requery phoneme sequence of word (level Text) 'beautiful'

sl1=query('ae','Text=beautiful')
requery_hier(sl1,level='Phoneme')

## Upward requery words (level Text) of Phonetic p elements

sl1=query('ae','Phonetic=p')
requery_hier(sl1,level='Text')

## Combined requery: last phonemes of words beginning with 'an'

sl1=query('ae',"Text=~an.*")
requery_seq(requery_hier(sl1,level='Phoneme'),offsetRef = 'END')

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}}
}
\keyword{EQL2}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

