% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/requery.database.R
\name{requery_hier}
\alias{requery_hier}
\title{Requery hierarchical context of segment list}
\usage{
requery_hier(seglist, level = NULL, resultType = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on (type: 'emuRsegs')}

\item{level}{character string: result level}

\item{resultType}{type (class name) of result (for now only 'emuRsegs')}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class resultType (default: 'emuRsegs')
}
\description{
WARNING! Experimental, syntax and semantics may change!!
}
\examples{
\dontrun{

## Requery phoneme sequence of word (Text) 'beautiful'
sl1=query('ae','Text=beautiful')
requery_hier(sl1,level='Phoneme')


}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}}
}
\keyword{EQL2}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

