% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requery.database.R
\name{requery_hier}
\alias{requery_hier}
\title{Requery hierarchical context of segment list}
\usage{
requery_hier(seglist, level = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on (type: \link{emuRsegs})}

\item{level}{character string: result level}

\item{dbUUID}{optional UUID odf emuDB}
}
\value{
result set object of class \link{emuRsegs}
}
\description{
Requery hierarchical context of segment list
}
\details{
For each segment of the input list the function looks for hierarchically connected items of start and end item for the result segment on the given target level.
      For the start item the item with the lowest sample position is chosen, for the end item with the highest sample position.
      The result list will have the same order as the input list. If result and the input list have the same length, the result segments have the same position as the corresponding ones in the input list. 
      If the length of the lists differ a synchronous ordering is not possible and therefore a warning is generated.
}
\examples{
\dontrun{

## Downward requery phoneme sequence of word (level Text) 'beautiful'

sl1=query('ae','Text=beautiful')
requery_hier(sl1,level='Phoneme')

## Upward requery words (level Text) of Phonetic p elements

sl1=query('ae','Phonetic=p')
requery_hier(sl1,level='Text')

## Combined requery: last phonemes of words beginning with 'an'

sl1=query('ae',"Text=~an.*")
requery_seq(requery_hier(sl1,level='Phoneme'),offsetRef = 'END')

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}} \code{\link{emuRsegs}}
}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

