% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requery.database.R
\name{requery_hier}
\alias{requery_hier}
\title{Requery hierarchical context of a segment list in an emuDB}
\usage{
requery_hier(seglist, level = NULL, dbUUID = NULL)
}
\arguments{
\item{seglist}{segment list to requery on (type: \link{emuRsegs})}

\item{level}{character string: name of target level}

\item{dbUUID}{optional UUID of emuDB, if the emuDB name in the input segment list is not unique}
}
\value{
result set object of class \link{emuRsegs}
}
\description{
Function to requery hierarchical context of a segment list queried from an emuDB
}
\details{
A segment is defined as a single item or a chain of items from the respective level, e.g. if a level in a bundle instance has labels 'a', 'b' and 'c' in that order, 'a' or 'a->b' oder 'a->b->c' are all valid segments, but not 'a->c'.
For each segment of the input segment list \code{seglist} the function checks the start and end item for hierarchically linked items in the given target level, and based on them constructs segments in the target level.
As the start item in the resulting segment the item with the lowest sample position is chosen; for the end item that with the highest sample position.
If result and input segment list have the same length (for each input segment one segment on the target level was found), the result segment list has the same length and order as the input list; 
in 'upwards' requeries this can cause a resulting segment list to contain two (or more) copies of the same segment, if the same item from the input list was linked twice or more to an item of the target level, e.g. a phoneme 'p' requeried to the word level might result in two identical segments 'Papa' in the result list. 
If the length of input and output list differ (e.g. because a link is missing in the emuDB), a synchronous ordering is not possible and therefore an error is generated.
}
\examples{
\dontrun{

## Downward requery: find 'Phoneme' sequences of all words 'beautiful' (of level 'Text')
## Note that the resulting segments consists of phoneme sequences and have therefore 
## the same length as the word segments.

sl1=query('ae','Text=beautiful')
requery_hier(sl1,level='Phoneme')

## Upward requery: find all word segments that dominate a 'p' on level 'Phoneme'
## Note that the resulting segments are larger than the input segments,
## because they contain the complete words.

sl1=query('ae','Phonetic==p')
wl1=requery_hier(sl1,level='Text')
wl1

## Why is there a 'p' the word 'emphazised'? Requery the whole words back down to 'Phoneme' level:

requery_hier(wl1,level='Phoneme')

## ... because of 'stop epenthesis' a 'p' is inserted between 'm' and 'f'

## Combined requery: last phonemes of all words beginning with 'an'.
## Note that we use a regular expression 'an.*' (EQL operator '=~') in the query.

sl1=query('ae',"Text=~an.*")
requery_seq(requery_hier(sl1,level='Phoneme'),offsetRef = 'END')

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{query}} \code{\link{requery_seq}} \code{\link{emuRsegs}}
}
\keyword{database}
\keyword{emuDB}
\keyword{requery}

