% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_all}
\alias{runBASwebservice_all}
\title{Runs several BAS webservices with default options, starting from an orthographic transcription}
\usage{
runBASwebservice_all(handle, transcriptionLabel, language, orthoLabel = "ORT",
  canoLabel = "KAN", mausLabel = "MAU", minniLabel = "MINNI",
  sylLabel = "MAS", canoSylLabel = "KAS", chunkLabel = "TRN",
  resume = FALSE, verbose = TRUE)
}
\arguments{
\item{handle}{emuDB handle}

\item{transcriptionLabel}{name of the label (not level!) containing an orthographic transcription.}

\item{language}{language(s) to be used. If you pass a single string (e.g. "deu-DE"), this language will be used for all bundles.
Alternatively, you can select the language for every bundle individually. To do so, you must pass a data frame with the columns
session, bundle, language. This data frame must contain one row for every bundle in your emuDB. 
Up-to-date lists of the languages accepted by all webservices can be found here:
\url{https://clarin.phonetik.uni-muenchen.de/BASWebServices/services/help}}

\item{orthoLabel}{label name for orthographic words}

\item{canoLabel}{label name for canonical pronunciations of words}

\item{mausLabel}{label name for the MAUS segmentation}

\item{minniLabel}{label name for the MINNI segmentation}

\item{sylLabel}{label name for syllable segmentation}

\item{canoSylLabel}{label name for syllabified canonical pronunciations of words}

\item{chunkLabel}{label name for the chunk segmentation
Please note that the chunk segmentation will only be generated if your emuDB contains
audio files beyond the one minute mark.}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue
that caused the error), you can set resume=TRUE to recover any progress made up to that point. This
will only work if you have not run any other emuR functions (such as the query function or other
webservice functions) in the meantime, as they are likely to delete your temporary data.}

\item{verbose}{Display progress bars and other information}
}
\description{
This function calls the BAS webservices G2P, MAUS, Pho2Syl, MINNI and (if necessary) Chunker.
Starting from an orthographic transcription, it derives a tokenized orthographical word tier
using the G2P tool. It also derives canonical pronunciations (in SAMPA) for the words.
If the audio file is longer than 60 seconds, the function then calls the Chunker webservice
to presegment the recording. Subsequently, the webservice MAUS is called to derive a phonetic
segmentation. A second, rough segmentation is created by running the phoneme decoder MINNI.
Finally, syllabification is performed by calling Pho2Syl. All necessary level, attribute and link
definitions are created in the process. \strong{This function requires an internet connection.}
}
\details{
Note that this function will run all BAS webservices in default mode. If you wish to change
options, you must use the individual runBASwebservices functions. This will also allow you to carry
out manual corrections in between the individual steps, or to use different languages for different
webservices.
}
\seealso{
Other BAS.webservice.functions: \code{\link{runBASwebservice_chunker}},
  \code{\link{runBASwebservice_g2pForPronunciation}},
  \code{\link{runBASwebservice_g2pForTokenization}},
  \code{\link{runBASwebservice_maus}},
  \code{\link{runBASwebservice_minni}},
  \code{\link{runBASwebservice_pho2sylCanonical}},
  \code{\link{runBASwebservice_pho2sylSegmental}}
}

