% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bas_webservices.R
\name{runBASwebservice_pho2sylCanonical}
\alias{runBASwebservice_pho2sylCanonical}
\title{Adds syllabified word labels to a word level that already contains a canonical pronunciation label.}
\usage{
runBASwebservice_pho2sylCanonical(handle, canoLabel, language,
  canoSylLabel = "KAS", params = list(), verbose = TRUE, resume = FALSE)
}
\arguments{
\item{handle}{emuDB handle}

\item{canoLabel}{name of the label (not level!) containing a canonical pronunciation of the words.}

\item{language}{language to be used. Up-to-date lists of the languages accepted by all webservices can be found here:
\url{https://clarin.phonetik.uni-muenchen.de/BASWebServices/services/help}}

\item{canoSylLabel}{label name for syllabified canonical pronunciations of words}

\item{params}{named list of parameters to be passed on to the webservice. It is your own reponsibility to
ensure that these parameters are compatible with the webservice API
(see \url{https://clarin.phonetik.uni-muenchen.de/BASWebServices/services/help}).
Some options accepted by the API (e.g. output format) cannot be set when calling a webservice from within emuR,
and will be overridden.}

\item{verbose}{Display progress bars and other information}

\item{resume}{If a previous call to this function has failed (and you think you have fixed the issue
that caused the error), you can set resume=TRUE to recover any progress made up to that point. This
will only work if you have not run any other emuR functions (such as the query function or other
webservice functions) in the meantime, as they are likely to delete your temporary data.}
}
\description{
Adds syllabified word labels to a word level that already contains a canonical pronunciation label.
}
\seealso{
Other BAS.webservice.functions: \code{\link{runBASwebservice_all}},
  \code{\link{runBASwebservice_chunker}},
  \code{\link{runBASwebservice_g2pForPronunciation}},
  \code{\link{runBASwebservice_g2pForTokenization}},
  \code{\link{runBASwebservice_maus}},
  \code{\link{runBASwebservice_minni}},
  \code{\link{runBASwebservice_pho2sylSegmental}}
}

