% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-objDocs.R
\name{segmentlist}
\alias{segmentlist}
\alias{emusegs}
\title{Segment list}
\format{multi-columned matrix one row per segment 
\itemize{ 
  \item columnlabel 
  \item columnsegment onset time 
  \item columnsegment offset time 
  \item columnutterance name 
}}
\description{
A segment list is the result type of legacy Emu query.
}
\examples{

   data(demo.vowels)
   
   #demo.vowels is a segment list
   demo.vowels

}
\seealso{
\code{\link{query}}, \code{\link{demo.vowels}}
}
\keyword{classes}
