% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.DBconfig.EMUwebAppConfig.R
\name{SetGetlevelCanvasesOrder}
\alias{SetGetlevelCanvasesOrder}
\alias{get_levelCanvasesOrder}
\alias{set_levelCanvasesOrder}
\title{Set / Get level canvases order of emuDB}
\usage{
set_levelCanvasesOrder(dbName, perspectiveName, order, dbUUID = NULL)

get_levelCanvasesOrder(dbName, perspectiveName, dbUUID = NULL)
}
\arguments{
\item{dbName}{name of loaded emuDB}

\item{perspectiveName}{name of perspective}

\item{order}{character vector containig names of levelDefinitions}

\item{dbUUID}{optional UUID of loaded emuDB}
}
\description{
Set / Get level canvases order of emuDB. Level canvases refer to levels of 
the type "SEGMENT" or "EVENT" that are displayed by the EMU-webApp. Levels 
of type "ITEM" can always be displayed using the show hierarchy modal of the
web application but may not be displayed as level canvases. 
For more information on the structural elements of an emuDB 
see \code{vignette{emuDB}}.
}
\examples{
\dontrun{

##################################
# prerequisite: loaded "ae" emuDB 
# (see ?load_emuDB for more information)

# get level canvases order of "ae" emuDB
order = get_levelCanvasesOrder(dbName = "ae",
                               perspectiveName = "default")

# reverse the level canvases order of "ae" emuDB
set_levelCanvasesOrder(dbName = "ae"
                       perspectiveName = "default",
                       order = rev(order))
                       
# get level canvases order of "ae" emuDB                       
get_levelCanvasesOrder(dbName = "ae",
                       perspectiveName = "default")
}

}
\keyword{DBconfig}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

