% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sortmatrix}
\alias{sortmatrix}
\title{Sort matrix by label}
\usage{
sortmatrix(mat, labs = dimnames(mat)[[2]])
}
\arguments{
\item{mat}{A mu+ sement matrix.}

\item{labs}{A label vector which has the same number of columns as
\code{mat}.}
}
\value{
Returns a sorted matrix by label, created from \code{mat}.
}
\description{
Sorts matrix by label
}
\seealso{
label, phon
}
\keyword{misc}

