% Generated by roxygen2 (4.0.1): do not edit by hand
\name{store.database}
\alias{store.database}
\title{Store EMU database to directory}
\usage{
store.database(db, targetDir, rewriteSSFFTracks = TRUE, showProgress = TRUE)
}
\arguments{
\item{db}{EMU database (in R workspace)}

\item{targetDir}{target directory}

\item{rewriteSSFFTracks}{if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC)}

\item{showProgress}{show progress bar}
}
\description{
Store EMU database to directory
}
\examples{
\dontrun{
# Store database object ae to directory /homes/mylogin/EMUnew/

store.database(ae,"/homes/mylogin/EMUnew/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.database}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuSX}

