% Generated by roxygen2 (4.0.2): do not edit by hand
\name{store.emuDB}
\alias{store.emuDB}
\title{Store EMU database to directory}
\usage{
store.emuDB(db, targetDir, options = list(), showProgress = TRUE)
}
\arguments{
\item{db}{EMU database (in R workspace)}

\item{targetDir}{target directory}

\item{options}{list of options}

\item{showProgress}{show progress bar}
}
\description{
options is a list of key value pairs:
rewriteSSFFTracks if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: TRUE
}
\examples{
\dontrun{
# Store database object ae to directory /homes/mylogin/EMUnew/

store.emuDB(ae,"/homes/mylogin/EMUnew/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load.emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

