% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{store}
\alias{store}
\title{Store EMU database to directory}
\usage{
store(dbName = NULL, targetDir, dbUUID = NULL, options = NULL,
  showProgress = TRUE)
}
\arguments{
\item{dbName}{name of emuDB}

\item{targetDir}{target directory}

\item{dbUUID}{optional: UUID of database}

\item{options}{list of options}

\item{showProgress}{show progress bar}
}
\description{
Store EMU database to directory
}
\details{
options is a list of key value pairs:
rewriteSSFFTracks if TRUE rewrite SSF tracks instead of file copy to get rid of big endian encoded SSFF files (SPARC), default: FALSE
ignoreMissingSSFFTrackFiles if TRUE missing SSFF track files are ignored, default: FALSE
symbolicLinkSignalFiles if TRUE signal files are symbolic linked instead of copied. Implies: rewriteSSFFTracks=FALSE, Default: FALSE
}
\examples{
\dontrun{
# Store database 'ae' to directory /homes/mylogin/EMUnew/

  store('ae',"/homes/mylogin/EmuStore/")

}
}
\author{
Klaus Jaensch
}
\seealso{
\code{\link{load_emuDB}}
}
\keyword{Emu}
\keyword{database}
\keyword{emuDB}

