% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-emuDBhandle.R
\name{summary.emuDBhandle}
\alias{summary.emuDBhandle}
\title{Print summary of loaded EMU database (emuDB).}
\usage{
summary.emuDBhandle(object, ...)
}
\arguments{
\item{dbName}{name of emuDB}

\item{dbUUID}{optional UUID of emuDB}
}
\description{
Gives an overview of an EMU database.
Prints database name, UUID, base directory path, session and bundle count and informations about signal track, annotation level, attribute and link definitions.
}

