% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emusegs.R
\name{summary.emusegs}
\alias{summary.emusegs}
\title{summary emusegs}
\usage{
\method{summary}{emusegs}(object, ...)
}
\arguments{
\item{object}{the segmentlist}

\item{\dots}{nothing special}
}
\description{
summarizes data in emu segment lists
}
\keyword{internal}

