% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdata.class.R
\name{summary.trackdata}
\alias{summary.trackdata}
\title{summary trackdata}
\usage{
\method{summary}{trackdata}(object, ...)
}
\arguments{
\item{object}{track data object}

\item{\dots}{see summary}
}
\description{
summarizes trackdata objects
}
\keyword{internal}
