% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-annotations_crud.R
\name{update_itemsInLevel}
\alias{update_itemsInLevel}
\title{Change item labels programmatically}
\usage{
update_itemsInLevel(emuDBhandle, itemsToUpdate, rewriteAllAnnots = TRUE,
  verbose = TRUE)
}
\arguments{
\item{emuDBhandle}{emuDB handle as returned by \code{\link{load_emuDB}}}

\item{itemsToUpdate}{A data frame with the columns
\itemize{
\item\code{session},
\item\code{bundle},
\item\code{level},
\item\code{sequenceIndex},
\item\code{attribute}, and
\item \code{label}.
}
*None* of the columns should be factors.
\code{sequenceIndex} must be numeric (natural-valued), all other columns
must be of type character.}

\item{rewriteAllAnnots}{should changes be written to file system (_annot.json
files) (intended for expert use only)}

\item{verbose}{if set to \code{TRUE}, more status messages are printed}
}
\description{
Change item labels programmatically
}
