% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax.dft.5}
\alias{vowlax.dft.5}
\title{Spectral matrix centred at the temporal midpoint of the vowels from the
segment list vowlax.}
\format{spectral matrix}
\description{
An EMU dataset
}
\keyword{datasets}
