% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{vowlax.fund.5}
\alias{vowlax.fund.5}
\title{Vector of fundamental frequency extracted at the temporal midpoint from the
segment list vowlax.}
\format{vector of fundamental frequency}
\description{
An EMU dataset
}
\keyword{datasets}

