% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax}
\alias{vowlax}
\title{Segment list of four lax vowels, read speech, one male and one female
speaker of Standard North German from database kielread.}
\format{segmentlist}
\description{
An EMU dataset
}
\keyword{datasets}

