% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax.right}
\alias{vowlax.right}
\title{Vector of labels following the vowels from the segment list vowlax}
\format{vector of phoenme labels}
\description{
An EMU dataset
}
\keyword{datasets}

