% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emuR-dataDocs.R
\docType{data}
\name{vowlax.rms.5}
\alias{vowlax.rms.5}
\title{Vector of RMS energy values at the temporal midpoint extracted at the
temporal midpoint from the segment list vowlax}
\format{vector of RMS energy values}
\description{
An EMU dataset
}
\keyword{datasets}

