% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR-bundleList.R
\name{write_bundleList}
\alias{write_bundleList}
\title{write bundleList}
\usage{
write_bundleList(
  emuDBhandle,
  name,
  bundleList,
  seglist,
  updateDBconfig = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{emuDBhandle}{emuDB handle object (see \link{load_emuDB})}

\item{name}{name of bundleList (excluding the _bundleList.json suffix)}

\item{bundleList}{tibble/data.frame with the columns \code{session}, \code{name}, 
\code{comment} (optional), \code{finishedEditing} (optional). Use \link{list_bundles}}

\item{seglist}{segment list returned by \link{query} function. If set the 
\code{bundleList} parameter will be ignored and a bundleList will be created by
collapsing the segments as timeAnchors into the \code{_bundleList.json}}

\item{updateDBconfig}{if set to TRUE (the default) DBconfig will be updated 
with the fields}

\item{verbose}{be verbose
\itemize{
 \item \code{"bundleComments": true}
 \item \code{"bundleFinishedEditing": true}
}}
}
\description{
write bundleList JSON file to emuDB
}
\details{
Write bundleList JSON file to emuDB sub-dir \code{bundleLists/}
}
